% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHA.R
\name{AHA}
\alias{AHA}
\title{All-H Analyzer}
\usage{
AHA(SOM, ngen = 100, silent = FALSE)
}
\arguments{
\item{SOM}{An object of class \linkS4class{SOM}}

\item{ngen}{Integer, the number of generations for which to run the simulation}

\item{silent}{Logical, indicates whether to silence messages to the R console}
}
\value{
A named list containing vectors of state variables (by simulation and generation). See \linkS4class{SMSE} object description.
}
\description{
Wrapper function for an implementation of All-H Analyzer (\href{https://www.streamnet.org/home/data-maps/hatchery-reform/hsrg-tools/}{AHA}) in R. Can be used to compare outputs between AHA
and salmonMSE.
}
\references{
Hatchery Scientific Review Group. 2020. All-H Analyzer Tool Guide and Documentation. May 2020.
}
