% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_func.R
\name{brood_func}
\alias{brood_func}
\title{Brood function}
\usage{
brood_func(
  Nage_NOS,
  Nage_HOS,
  stray_external,
  m,
  m_stray,
  s_enroute,
  hatchery_args
)
}
\arguments{
\item{Nage_NOS}{Array \verb{[nage, n_g]} of natural-origin fish}

\item{Nage_HOS}{Array \verb{[nage, n_r]} of hatchery-origin fish}

\item{stray_external}{Array \verb{[nage, n_r]} of hatchery-origin strays}

\item{m}{Numeric, mark rate of \code{Nage_HOS}}

\item{m_stray}{Numeric, mark rate of \code{stray_external}}

\item{s_enroute}{Numeric, en-route survival}

\item{hatchery_args}{List of various hatchery arguments created by \code{\link[=define_hatchery_args]{define_hatchery_args()}} and adjusted by \code{\link[=ProjectSOM]{ProjectSOM()}}.}
}
\value{
Named list:
\itemize{
\item \code{broodtake} list returned by \code{\link[=calc_broodtake]{calc_broodtake()}}
\item \code{hatchery_production} list returned by \code{\link[=calc_yearling]{calc_yearling()}}
\item \code{spawners} list returned by \code{\link[=calc_spawners]{calc_spawners()}}
}
}
\description{
Calculates broodtake and in-river removals from escapement of marine fisheries. This function also applies en-route mortality.
}
\seealso{
\code{\link[=calc_broodtake]{calc_broodtake()}}
}
\keyword{internal}
