% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchery.R
\name{calc_broodtake}
\alias{calc_broodtake}
\alias{calc_broodtake_custom}
\alias{.broodtake_func}
\title{Solve for broodtake numbers}
\usage{
calc_broodtake(
  NO,
  HO,
  stray,
  brood_import,
  ptarget_NOB,
  pmax_NOB,
  phatchery,
  egg_target,
  p_female,
  fec,
  s_prespawn,
  m
)

calc_broodtake_custom(
  f_brood,
  NO,
  HO,
  stray,
  p_female,
  fec,
  s_prespawn,
  m,
  egg_target
)

.broodtake_func(
  ptake_unmarked,
  NO,
  HO,
  stray,
  brood_import,
  phatchery,
  p_female,
  fec,
  egg_target,
  s_prespawn,
  ptarget_NOB,
  m = 1,
  opt = TRUE
)
}
\arguments{
\item{NO}{Matrix \verb{[nage, n_g]}, natural-origin fish available for broodtake}

\item{HO}{Matrix \verb{[nage, n_r]}, hatchery-origin fish, potentially available for broodtake}

\item{stray}{Matrix \verb{[nage, n_r]}, hatchery-origin strays available for broodtake}

\item{brood_import}{Vector \verb{[nage]} of imported brood}

\item{ptarget_NOB}{Numeric, target proportion of \code{NOB/(NOB + HOB)}. If \code{m = 1}, then the realized pNOB should be ptarget_NOB.
If \code{m < 1}, then the system achieves \verb{unmarked brood/(NOB + HOB)  = ptarget_NOB}.}

\item{pmax_NOB}{Numeric, maximum proportion of \code{NOB/NO}}

\item{phatchery}{Numeric, proportion of \code{HO} that return to hatchery instead of the spawning ground. HOB is taken from this subset. Set to \code{NA} to
obtain HOB on the way to spawning ground.}

\item{egg_target}{Numeric, target egg production from which to back-calculate brood numbers}

\item{p_female}{Vector \verb{[nage]} of proportion female to calculate egg production}

\item{fec}{Vector \verb{[nage]}, egg production per female}

\item{s_prespawn}{Numeric, survival of brood prior to egg production}

\item{m}{Numeric, mark rate of \code{HO} used. Discounts \code{ptarget_NOB} and \code{pmax_NOB} based on marked fish}

\item{f_brood}{Function that calculates the brood numbers}

\item{ptake_unmarked}{Numeric, proportion of unmarked fish used for brood}

\item{opt}{Logical, whether the function is used for optimization (TRUE) or reporting (FALSE)}
}
\value{
\code{calc_broodtake()} returns a list from \code{.broodtake_func()}

\code{calc_broodtake_custom()} returns a named list, same format as \code{calc_broodtake()}

\code{.broodtake_func()} returns a numeric if \code{opt = TRUE}: \code{log(p_unmarked) - log(ptarget_NOB)}, otherwise
a named list of brood and egg production.
\itemize{
\item \code{egg_NOB} Matrix \verb{[nage, n_g]}
\item \code{egg_HOB_unmarked} Matrix \verb{[nage, n_r]} (including strays)
\item \code{egg_HOB_marked} Matrix \verb{[nage, n_r]}
\item \code{egg_HOB_import} Vector \verb{[nage]}
\item \code{ptake_unmarked} Numeric
\item \code{ptake_marked} Numeric
\item \code{pNOB} Numeric
\item \code{NOB} Numeric
\item \code{HOB_unmarked} Matrix \verb{[nage, n_r]}
\item \code{HOB_marked} Matrix \verb{[nage, n_r]}
\item \code{HOB_import} Matrix \verb{[nage, n_r]}
\item \code{HOB_stray} Vector \verb{[nage]}
}
}
\description{
Internal functions to calculate broodtake with various constraints, called by \code{\link[=brood_func]{brood_func()}}.
\code{calc_broodtake()} is a wrapper function.

\code{calc_broodtake_custom()} uses a user-provided function to generate brood numbers and adjusts
downwards if egg production exceeds the target.

\code{.broodtake_func()} is the optimization function used to calculate broodtake
}
\keyword{internal}
