% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{calc_ref}
\alias{calc_ref}
\title{Reference points}
\usage{
calc_ref(SOM, rel_F, check = TRUE, maximize = c("MSY", "MER"))
}
\arguments{
\item{SOM}{An object of class \linkS4class{SOM}}

\item{rel_F}{Numeric length 2, indicates the relative effort in the preterminal and terminal fisheries, with a maximum value of 1.
For example, \code{c(1, 0)} indicates a yield calculation for only the preterminal fishery. By default, uses the ratio of the harvest rates in the operating model.}

\item{check}{Logical, whether to check the SOM object using \code{\link[=check_SOM]{check_SOM()}}}

\item{maximize}{Character, whether the MSY calculation is the optimum that maximizes catch (\code{"MSY"}) or excess recruitment (\code{"MER"})}
}
\value{
Matrix of various state variables (catch, exploitation rate, egg production, spawners) at MSY and Sgen by simulation
}
\description{
Calculate MSY and Sgen reference points for the operating model. Uses the biological parameters (maturity, natural mortality)
in the last year of the projection.
}
