% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchery.R
\name{calc_spawners}
\alias{calc_spawners}
\title{Calculate spawners after broodtake}
\usage{
calc_spawners(broodtake, NO, HO, stray, phatchery, premove_HOS, premove_NOS, m)
}
\arguments{
\item{broodtake}{List, returned by \code{\link[=calc_broodtake]{calc_broodtake()}}}

\item{NO}{Matrix \verb{[nage, n_g]}, natural-origin in-river return}

\item{HO}{Matrix \verb{[nage, n_r]}, hatchery-origin in-river return}

\item{stray}{Matrix \verb{[nage, n_r]}, hatchery-origin strays in-river return}

\item{phatchery}{Numeric, proportion of \code{HO} that return to hatchery instead of the spawning ground. HOB is taken from this subset. Set to \code{NA} to
obtain HOB on the way to spawning ground.}

\item{premove_HOS}{Numeric, proportion of HO fish to remove (after brood). Can also be a function}

\item{premove_NOS}{Numeric, proportion of NO fish to remove (after brood). Can also be a function}

\item{m}{Numeric, mark rate}
}
\value{
Named list:
\itemize{
\item NOS \verb{[nage, n_g]} natural-origin spawners
\item HOS \verb{[nage, n_r]} hatchery-origin spawners
\item HOS_stray \verb{[nage, n_r]} spawners that are strays
\item NO_remove \verb{[nage, n_g]} natural-origin fish removed before spawning
\item HO_remove \verb{[nage, n_r]} hatchery-origin fish removed before spawning
}
}
\description{
Internal function that calculates remaining spawners after broodtake and additional removals.
}
\keyword{internal}
