% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchery.R
\name{calc_yearling}
\alias{calc_yearling}
\alias{.yearling_func}
\title{Calculate hatchery releases}
\usage{
calc_yearling(egg_target, s_yearling, s_subyearling, p_yearling, p_subyearling)

.yearling_func(
  p_egg_yearling,
  egg_target,
  s_yearling,
  s_subyearling,
  p_yearling,
  opt = TRUE
)
}
\arguments{
\item{egg_target}{Numeric}

\item{s_yearling}{Numeric, survival from egg to yearling}

\item{s_subyearling}{Numeric, survival from egg to subyearling}

\item{p_yearling}{Vector \verb{[n_r]}, proportion of releases at subyearling stage, where \code{sum(p_yearling, p_subyearling) = 1}}

\item{p_egg_yearling}{Numeric, proportion to eggs to become yearlings}

\item{opt}{Logical, whether the function is used for optimization (TRUE) or reporting (FALSE)}
}
\value{
\code{calc_yearling()} returns a named list returned by \code{.yearling_func()}

\code{.yearling_func()} returns a numeric if \code{opt = TRUE}: \code{yearling/(yearling + subyearling) - p_yearling}.
Otherwise, log(p_unmarked) - log(ptarget_NOB)`, otherwise a named list:
\itemize{
\item \code{yearling} Vector \verb{[n_r]} of yearling releases
\item \code{subyearling} Vector \verb{[n_r]} of subyearling releases
}
}
\description{
From egg target, calculate production of yearlings and subyearlings conditional on survival and
proportion of releases.

\code{.yearling_func()} is the optimization function that determines proportions.
}
\keyword{internal}
