% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_func.R
\name{catch_func}
\alias{catch_func}
\title{Catch function}
\usage{
catch_func(
  NO,
  HO,
  type = c("u", "catch"),
  U,
  K,
  V,
  ForeErr = 1,
  ImpErr = 1,
  MSF = FALSE,
  m = 1,
  release_mort = 0,
  p_mature_NO = array(1, dim(NO)),
  p_mature_HO = array(1, dim(HO)),
  AEQ_NO = array(1, dim(NO)),
  AEQ_HO = array(1, dim(HO))
)
}
\arguments{
\item{NO}{Array \verb{[ns, nage, n_g]} of natural-origin fish}

\item{HO}{Array \verb{[ns, nage, n_r]} of hatchery-origin fish}

\item{type}{Character. Whether to calculate removals from harvest rate \code{"u"}or catch \code{"catch"}}

\item{U}{Numeric or function. Harvest rate of fishery}

\item{K}{Numeric or function. Total catch of the fishery}

\item{V}{Matrix \verb{[ns, nage]} Relative vulnerability by age class to fishery}

\item{ForeErr}{Numeric. Multiplicative forecast error for control rule, only used if either \code{U} or \code{K} is a function (1 = no error)}

\item{ImpErr}{Numeric. Implementation error of harvest control (0 = no error)}

\item{MSF}{Logical, whether fishing is mark-selective}

\item{m}{Numeric vector length \verb{[ns]}, mark rate of hatchery-origin fish. Only used if \code{MSF = TRUE}.}

\item{release_mort}{Numeric vector length \verb{[ns]}, proportion of released unmarked fish that die. Only used if \code{MSF = TRUE}.}

\item{p_mature_NO}{Optional, array \verb{[ns, nage, n_g]} of maturity for natural-origin fish. Only use to calculate adult equivalents escapement.}

\item{p_mature_HO}{Optional, array \verb{[ns, nage, n_r]} of maturity for hatchery-origin fish. Only use to calculate adult equivalents escapement.}

\item{AEQ_NO}{Optional, array \verb{[ns, nage, n_g]} of adult equivalents for natural-origin fish.}

\item{AEQ_HO}{Optional, array \verb{[ns, nage, n_r]} of adult equivalents for hatchery-origin fish.}
}
\value{
Named list:
\itemize{
\item \code{K_NO} kept natural-origin catch, same dimension as \code{NO}
\item \code{K_HO} kept hatchery-origin catch, same dimension as \code{HO}
\item \code{D_NO} discarded (live + dead) natural-origin catch, same dimension as \code{NO}
\item \code{D_HO} discarded (live + dead) hatchery-origin catch, same dimension as \code{HO}
\item \code{DD_NO} dead discarded natural-origin catch, same dimension as \code{NO}
\item \code{DD_HO} dead discarded hatchery-origin catch, same dimension as \code{HO}
\item \code{Nsurv_NO} natural-origin survivors, same dimension as \code{NO}
\item \code{Nsurv_HO} hatchery-origin survivors, same dimension as \code{HO}
\item \code{U_NO} natural-origin harvest rate (ratio of kept catch and abundance), vector \code{ns}
\item \code{U_HO} hatchery-origin harvest rate (ratio of dead catch and abundance), vector \code{ns}
\item \code{Ex_NO} natural-origin exploitation rate (ratio of dead catch and abundance), vector \code{ns}
\item \code{Ex_HO} hatchery-origin exploitation rate (ratio of dead catch and abundance), vector \code{ns}
}
}
\description{
Internal function that does catch calculations for the preterminal and terminal marine fisheries.
}
\keyword{internal}
