% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_int.R
\name{CM_int}
\alias{CM_int}
\title{Estimation function for conditioning model}
\usage{
CM_int(p, d)
}
\arguments{
\item{p}{List of parameter variables. See \code{\link[=fit_CM]{fit_CM()}}.}

\item{d}{List of data variables. See \code{\link[=fit_CM]{fit_CM()}}.}
}
\value{
Numeric, objective function value (log-posterior)
}
\description{
Population dynamics model of an age structured salmon population. Used with RTMB to estimate historical reconstruction from data.
}
\author{
Q. Huynh with Stan code provided by J. Korman and C. Walters
}
\keyword{internal}
