% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMfun.R
\name{CM_MSY}
\alias{CM_MSY}
\alias{CM_Sgen}
\title{Plot reference points from conditioning model}
\usage{
CM_MSY(
  report,
  d,
  year1 = 1,
  simple = FALSE,
  index = NULL,
  mean_bio = FALSE,
  type = c("spawner", "egg", "u"),
  maximize = c("MSY", "MER"),
  AEQ = TRUE,
  ncores = 1,
  na.rm = FALSE
)

CM_Sgen(
  report,
  d,
  year1 = 1,
  simple = FALSE,
  index = NULL,
  mean_bio = FALSE,
  ncores = 1,
  na.rm = FALSE
)
}
\arguments{
\item{report}{List, output of state variables from individual MCMC samples, obtained with \code{\link[=get_report]{get_report()}}}

\item{d}{List of data variables, obtained with \code{\link[=get_CMdata]{get_CMdata()}}}

\item{year1}{Numeric, first year of model}

\item{simple}{Logical, whether to use Ricker lambert equations for MSY reference points (TRUE) or age-structured
optimization (FALSE)}

\item{index}{Integer vector to subset years with which to calculate reference points. Can be used
to reduce computation or average biological parameters from a subset of years, see \code{mean_bio} argument.
If \code{NULL}, uses all years of model.}

\item{mean_bio}{Logical, whether to average the natural mortality and maturity parameters across
years indicated in \code{index}}

\item{type}{Character, the type of reference point to calculate}

\item{maximize}{Character, whether the numerical optimization maximizes catch (MSY) or excess recruitment (MER). For testing only,
should not impact results. Only used if \code{simple = FALSE}}

\item{AEQ}{Logical, whether to use adult equivalents when calculating MSY for preterminal fisheries#' Only used if \code{simple = FALSE}. Should always be \code{TRUE}}

\item{ncores}{Numeric, number of processors for parallel computation. Useful if calculating numerically for many MCMC samples}

\item{na.rm}{Logical, whether to exclude negative values from the median in figures}
}
\value{
ggplot object
}
\description{
Plots time series of MSY reference points. Not used in automated reporting as they can be
computationally expensive and individual case studies may require specific assumptions.
}
\seealso{
\code{\link[=CM_prod]{CM_prod()}} \code{\link[=CM_Srep]{CM_Srep()}} \code{\link[=.CM_MSY]{.CM_MSY()}}
}
