% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare scenarios in markdown}
\usage{
compare(
  SMSE_list,
  names,
  col_vec,
  filename = "SMSEcompare",
  dir = tempdir(),
  open_file = TRUE,
  render_args = list(),
  ...
)
}
\arguments{
\item{SMSE_list}{\linkS4class{SMSE} object}

\item{names}{Character vector \code{length(SMSE_list)} to label individual model runs}

\item{col_vec}{Character vector \code{length(SMSE_list)} for custom colour schemes for comparing across model scenarios in figures}

\item{filename}{Character string for the name of the markdown and HTML files.}

\item{dir}{The directory in which the markdown and HTML files will be saved.}

\item{open_file}{Logical, whether the HTML document is opened after it is rendered.}

\item{render_args}{List of arguments to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{...}{Additional arguments (not used)}
}
\value{
Returns invisibly the output of \code{\link[rmarkdown:render]{rmarkdown::render()}}, typically the path of the output file
}
\description{
Generate a markdown report for multiple model runs to compare scenarios. 3-5 is likely the ideal number of scenarios for comparison.
}
\seealso{
\code{\link[=report]{report()}}
}
