% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_SOM.R
\name{define_hatchery_args}
\alias{define_hatchery_args}
\alias{define_habitat_args}
\alias{define_fitness_args}
\alias{define_SRRpars}
\title{Create list of parameters}
\usage{
define_hatchery_args(SOM)

define_habitat_args(SOM)

define_fitness_args(SOM)

define_SRRpars(SOM)
}
\arguments{
\item{SOM}{\linkS4class{SOM} operating model object}
}
\value{
\code{define_hatchery_args()} returns a length \code{nstocks} of hatchery parameters, in-river removals, and behavior of hatchery-origin
fish in the freshwater environment (fecundity, reproductive success, etc.)

\code{define_habitat_args()} returns a list of \linkS4class{Habitat} objects.

\code{define_fitness_args()} returns a length \code{nstocks} of fitness parameters.

\code{define_SRRpars_args()} returns a length \code{nstocks}, each of which is a data frame of stock-recruit parameters.
}
\description{
Internal functions that convert the operating model inputs into a list or data frame of parameters to pass on to other
internal functions.
}
\keyword{internal}
