% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{Deprecated}
\alias{Deprecated}
\alias{PNI50}
\alias{PNI80}
\alias{WILD50}
\alias{SMSY85}
\alias{Sgen100}
\title{Deprecated functions}
\usage{
PNI50(...)

PNI80(...)

WILD50(...)

SMSY85(...)

Sgen100(...)
}
\arguments{
\item{...}{Same arguments as \code{\link[=P_PNI50]{P_PNI50()}} and similar functions}
}
\value{
A vector of probabilities
}
\description{
Deprecated performance metric functions. These functions are now replaced with functions of same name
but prefixed with \code{P_}.
}
\seealso{
\code{\link[=P_PNI50]{P_PNI50()}}
}
\keyword{internal}
