% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMfun.R
\name{.CM_ER}
\alias{.CM_ER}
\alias{.CM_ERage}
\title{Calculate exploitation rate in conditioning model}
\usage{
.CM_ER(
  report,
  type = c("PT", "T", "all"),
  r = 1,
  index_AEQ = NULL,
  brood = FALSE,
  simplify = TRUE
)

.CM_ERage(report, type = c("PT", "T"), brood = FALSE, simplify = TRUE)
}
\arguments{
\item{report}{List, output of state variables from individual MCMC samples, obtained with \code{\link[=get_report]{get_report()}}}

\item{type}{Character, indicates type of variable to plot}

\item{r}{Integer, the release strategy for the figure (only if \code{annual = FALSE})}

\item{index_AEQ}{Optional integer vector to identify years from which to borrow natural mortality and maturity
to calculate adult equivalents for incomplete brood years. Only used if \code{at_age = FALSE}.}

\item{brood}{Logical, whether to show results by brood year or return year (FALSE)}

\item{simplify}{Logical, will return a matrix or array if TRUE, otherwise a list of output by MCMC simulation}
}
\value{
An array if \code{simplify = TRUE}. Otherwise, a list.
}
\description{
Internal functions for calculating aggregate exploitation rate (\code{.CM_ER()}) or by age class (\code{.CM_ER()}).
}
\keyword{internal}
