% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMfun.R
\name{.CM_prod}
\alias{.CM_prod}
\alias{.CM_MSY}
\title{Reference points for conditioning model}
\usage{
.CM_prod(report, d, index = NULL, mean_bio = FALSE)

.CM_MSY(
  report,
  d,
  simple = TRUE,
  index = NULL,
  mean_bio = FALSE,
  type = c("spawner", "egg", "u", "Sgen"),
  AEQ = TRUE,
  maximize = c("MSY", "MER"),
  ncores = 1
)
}
\arguments{
\item{report}{List, output of state variables from individual MCMC samples, obtained with \code{\link[=get_report]{get_report()}}}

\item{d}{List of data variables, obtained with \code{\link[=get_CMdata]{get_CMdata()}}}

\item{index}{Integer vector to subset years with which to calculate reference points. Can be used
to reduce computation or average biological parameters from a subset of years, see \code{mean_bio} argument.
If \code{NULL}, uses all years of model.}

\item{mean_bio}{Logical, whether to average the natural mortality and maturity parameters across
years indicated in \code{index}}

\item{simple}{Logical, whether to use Ricker lambert equations for MSY reference points (TRUE) or age-structured
optimization (FALSE)}

\item{type}{Character, the type of reference point to calculate}

\item{AEQ}{Logical, whether to use adult equivalents when calculating MSY for preterminal fisheries#' Only used if \code{simple = FALSE}. Should always be \code{TRUE}}

\item{maximize}{Character, whether the numerical optimization maximizes catch (MSY) or excess recruitment (MER). For testing only,
should not impact results. Only used if \code{simple = FALSE}}

\item{ncores}{Numeric, number of processors for parallel computation. Useful if calculating numerically for many MCMC samples}
}
\value{
Matrix, dimension \verb{[length(index), length(report)]}. If \code{mean_bio = TRUE}, matrix has 1 row.
}
\description{
Internal functions that calculate productivity (\code{.CM_prod()}) and reference points (\code{.CM_MSY()})
from the conditioning model.
These functions can be used to calculate values for a subset of years (productivity can vary
in time with natural mortality and maturity), or from average biological parameters in a subset of years.
}
\seealso{
\code{\link[=CM_MSY]{CM_MSY()}}
}
\keyword{internal}
