% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchery.R
\name{.egg_func}
\alias{.egg_func}
\title{Egg production function}
\usage{
.egg_func(
  ptake = 1,
  N,
  gamma = 1,
  fec,
  p_female,
  s_prespawn,
  val = 0,
  opt = TRUE
)
}
\arguments{
\item{ptake}{Numeric, proportion of spawners that spawn}

\item{N}{Numeric, spawners}

\item{gamma}{Numeric, relative reproductive success of spawners}

\item{p_female}{Numeric, proportion female}

\item{s_prespawn}{Numeric, survival of spawners prior to egg production}

\item{val}{Numeric, target egg production. Used to optimize for \code{ptake} if \code{opt = TRUE}}

\item{opt}{Logical, whether the function is used to optimize for \code{ptake}}
}
\value{
Numeric
}
\description{
Simple wrapper function to calculate hatchery egg production
}
\keyword{internal}
