% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_func.R
\name{Effort_solver}
\alias{Effort_solver}
\title{Solve for fishing effort}
\usage{
Effort_solver(
  Eff,
  N,
  vul,
  ret,
  release_mort,
  type = c("u", "catch"),
  u = 0,
  K = 0,
  AEQ = array(1, dim(N)),
  p_mature = array(1, dim(N))
)
}
\arguments{
\item{Eff}{Numeric, fishing effort}

\item{N}{Array \verb{[ns, nage, n_r]}, total abundance (juvenile abundance for preterminal, return for terminal)}

\item{vul}{Array \verb{[ns, nage, n_r]}, fishery vulnerability}

\item{ret}{Vector \verb{[ns]}, retention rate}

\item{release_mort}{Vector \verb{[ns]}, release mortality as a proportion, between 0-1. Only relevant if \code{ret < 1}.}

\item{type}{Character, either \code{"catch"}, or \code{"u"}, whether to solve for kept catch or harvest rate, respectively}

\item{u}{Numeric, harvest rate target}

\item{K}{Numeric, kept catch target}

\item{AEQ}{Array \verb{[ns, nage, n_r]}, adult equivalents of catch}

\item{p_mature}{Array \verb{[ns, nage, n_r]}, proportion mature by age class (used to calculate adult equivalent escapement)}
}
\value{
Numeric. Returns the difference between the realized harvest rate (for the given value of \code{Eff}) and the target
}
\description{
Internal solver used by \code{catch_func()} to calculate the fishing effort needed to achieve target harvest rate or catch rate,
subject to partial retention due to mark-selective fishing.
Harvest rate is discounted by adult equivalents for preterminal fisheries.
}
\keyword{internal}
