% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enhancement.R
\name{Enhancement}
\alias{Enhancement}
\title{Hatchery Enhancement Evaluator}
\usage{
Enhancement(Pars = NULL, ...)
}
\arguments{
\item{Pars}{A list of key parameters}

\item{...}{Other arguments}
}
\value{
A list of answers
}
\description{
An example function.
}
\examples{
myresults <- Enhancement()
}
