% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_env.R
\docType{data}
\name{fitness_env}
\alias{fitness_env}
\title{Environment to store fitness calculates during the simulation}
\format{
An object of class \code{environment} of length 0.
}
\usage{
fitness_env
}
\description{
Environment to store fitness calculates during the simulation
}
\keyword{datasets}
