% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness_func}
\alias{fitness_func}
\alias{calc_zbar}
\title{Calculate mean phenotype and fitness}
\usage{
fitness_func(
  Egg_NOS,
  Egg_HOS,
  Egg_NOB,
  Egg_HOB,
  zbar_brood,
  fitness_args = list()
)

calc_zbar(
  Egg_NOS,
  Egg_HOS,
  Egg_NOB,
  Egg_HOB,
  zbar_prev,
  fitness_variance,
  theta,
  phenotype_variance,
  heritability
)
}
\arguments{
\item{Egg_NOS}{Numeric, egg production by natural-origin spawners}

\item{Egg_HOS}{Numeric, egg production by hatchery-origin brood}

\item{Egg_NOB}{Numeric, egg production by natural-origin brood}

\item{fitness_args}{List, containing \code{fitness_variance}, \code{theta}, \code{phenotype_variance}, \code{heritability}, \code{rel_loss}, and
\code{fitness_floor}}

\item{zbar_prev}{Numeric, mean phenotype of parents}

\item{fitness_variance}{Numeric, variance (omega-squared) of the fitness function}

\item{theta}{Numeric length 2, optimum phenotype value for the natural and hatchery environments, respectively.}

\item{phenotype_variance}{Numeric, variance (sigma-squared) of the phenotypic trait (theta)}

\item{heritability}{Numeric, heritability (h-squared) of the phenotypic trait}
}
\value{
\code{fitness_func} returns a list:
\itemize{
\item \code{zbar} Numeric length 2, mean phenotype of next generation to natural \code{zbar[1]} and hatchery \code{zbar[2]} environments
\item \code{fitness} Numeric length 2, fitness of next generation to natural and hatchery environments
\item \code{fitness_loss} Matrix \verb{[2, 3]} penalty in survival of next generation due to fitness effects
}

\code{calc_zbar} returns numeric length 2, mean phenotype of next generation
}
\description{
Quantitative genetics model of mean phenotypic trait and fitness of next brood year.

\code{calc_zbar()} is the internal function that calculates mean phenotype.
}
\keyword{internal}
