% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harvest_MMP.R
\name{get_F}
\alias{get_F}
\title{Calculate F from harvest rate}
\usage{
get_F(u = 0, M, ret = 0, release_mort = 0, Fmax = 20)
}
\arguments{
\item{u}{Harvest rate, between 0-1}

\item{M}{Instantaneous natural mortality rate}

\item{ret}{Retention rate, between 0-1}

\item{release_mort}{Release mortality as a proportion, between 0-1}

\item{Fmax}{Maximum allowable value of F}
}
\value{
Numeric
}
\description{
Solves for instantaneous fishing mortality rate from harvest rate
}
\keyword{internal}
