% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harvest_MMP.R
\name{Harvest_MMP}
\alias{Harvest_MMP}
\title{Harvest component of operating model}
\usage{
Harvest_MMP(
  x = 1,
  DataList,
  reps = 1,
  u_terminal,
  u_preterminal,
  m,
  release_mort,
  p_terminal = c(2, 5),
  p_preterminal = c(1, 4),
  ...
)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A nested list of \linkS4class{Data} objects by stock then fleet, generated by \code{\link[=multiMSE]{multiMSE()}}}

\item{reps}{The number of stochastic replicates to be returned by the function}

\item{u_terminal}{Harvest rate of retained catch in the terminal fishery}

\item{u_preterminal}{Harvest rate of retained catch in the pre-terminal fishery}

\item{m}{Mark rate of hatchery origin fish, as a proxy for fishery retention. Only used to calculate the fishing effort.
Retention in the operating model is specified in the \linkS4class{MOM} object}

\item{release_mort}{Release mortality of discarded fish. Only used to calculate the fishing effort.
Release mortality in the operating model is specified in the \linkS4class{MOM} object}

\item{p_terminal}{Population index for the recruitment that experiences the terminal fishing mortality}

\item{p_preterminal}{Population index for immature fish that experience the pre-terminal fishing mortality}

\item{...}{Not used}
}
\value{
A nested list of \linkS4class{Rec} objects, same dimension as \code{DataList}
}
\description{
A function used by openMSE to specify the fishing effort during the projections.
salmonMSE updates the arguments of this function from the salmon operating model.
}
\keyword{internal}
