% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harvest_MMP.R
\name{make_Harvest_MMP}
\alias{make_Harvest_MMP}
\title{Harvest component of operating model}
\usage{
make_Harvest_MMP(u)
}
\arguments{
\item{u}{Numeric. Harvest rate.}
}
\description{
A function that creates a multi-stock management procedure
}
