% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_func.R
\name{nextgen_SRR_func}
\alias{nextgen_SRR_func}
\alias{nextgen_habitat_func}
\title{Spawning and early life stage function}
\usage{
nextgen_SRR_func(
  Brood_Calcs,
  fec,
  p_female,
  hatchery_args = list(),
  fitness_args = list(),
  zbar_brood,
  SRRpars,
  p_LHG
)

nextgen_habitat_func(
  Brood_Calcs,
  Habitat,
  fec,
  p_female,
  hatchery_args = list(),
  fitness_args = list(),
  zbar_brood,
  p_LHG
)
}
\arguments{
\item{Brood_Calcs}{List, returned by \code{\link[=brood_func]{brood_func()}}}

\item{fec}{Vector \verb{[nage]}, fecundity at age of female fish}

\item{p_female}{Vector \verb{[nage]}, proportion female by age class}

\item{hatchery_args}{List of arguments created by \code{\link[=define_hatchery_args]{define_hatchery_args()}}}

\item{fitness_args}{List of arguments created by \code{\link[=define_fitness_args]{define_fitness_args()}}}

\item{Habitat}{\linkS4class{Habitat} object, modified by \code{\link[=ProjectSOM]{ProjectSOM()}}}
}
\value{
Named list:
\itemize{
\item \code{NOS} Matrix \verb{[nage, n_g]}
\item \code{HOS} Matrix \verb{[nage, n_r]}
\item \code{HOS_effective} Matrix \verb{[nage, n_r]}
\item \code{HOS_stray} Matrix \verb{[nage, n_r]}
\item \code{pHOSeff} Numeric
\item \code{pHOScensus} Numeric
\item \code{Egg_NOS} \verb{[nage, n_g]}
\item \code{Egg_HOS} \verb{[nage, n_r]}
\item \code{Smolt_RelOut} Vector \verb{[n_r]}
\item \code{Fry_NOS} Vector \verb{[n_g]}
\item \code{Fry_HOS} Vector \verb{[n_g]}
\item \code{Smolt_NOS} Vector \verb{[n_g]}
\item \code{Smolt_HOS} Vector \verb{[n_g]}
}
}
\description{
Calculates egg production from spawners arriving at spawning grounds, broodtake and in-river removals from escapement of marine fisheries. This function also applies en-route mortality.
}
\keyword{internal}
