% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{P_PNI50}
\alias{P_PNI50}
\alias{P_PNI80}
\alias{P_WILD50}
\alias{P_SMSY85}
\alias{P_SMSY80}
\alias{P_Sgen100}
\title{Example performance metrics}
\usage{
P_PNI50(SMSE, Ref = 0.5, Yrs = NULL)

P_PNI80(SMSE, Ref = 0.8, Yrs = NULL)

P_WILD50(SMSE, Ref = 0.5, Yrs = NULL)

P_SMSY85(SMSE, Ref = 0.85, Yrs = NULL)

P_SMSY80(SMSE, Ref = 0.85, Yrs = NULL)

P_Sgen100(SMSE, Ref = 1, Yrs = NULL)
}
\arguments{
\item{SMSE}{SMSE object returned by \code{\link[=salmonMSE]{salmonMSE()}}}

\item{Ref}{Threshold for the performance metric, used to calculate the probability that the metric exceeds this value}

\item{Yrs}{Numeric vector of length 2 to indicate the year range over which to summarize performance.
If NULL, the performance is summarized over all projection years.}
}
\value{
A vector of probabilities corresponding to population
}
\description{
Functions that evaluate return probabilities of outcomes from the simulations.
}
\details{
\itemize{
\item \code{P_PNI50} calculates the probability that PNI exceeds 0.50 (threshold for an integrated-transition population, Withler et al. 2018)
\item \code{P_PNI80} calculates the probability that PNI exceeds 0.80 (threshold for an integrated-wild population, Withler et al. 2018)
\item \code{P_WILD50} calculates the probability that at least 50 percent of natural spawners are wild
\item \code{P_SMSY80} calculates the probability that NOS exceeds 0.80 SMSY
\item \code{P_SMSY85} calculates the probability that NOS/SMSY exceeds 0.85 SMSY
\item \code{P_Sgen100} calculates the probability that NOS exceeds Sgen
}
}
\references{
Withler et al. 2018. Genetically Based Targets for Enhanced Contributions to Canadian Pacific Chinook Salmon Populations.
DFO Can. Sci. Advis. Sec. Res. Doc. 2018/019. xii + 88 p.
}
