% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE2SMSE.R, R/SOM2MOM.R
\name{MMSE2SMSE}
\alias{MMSE2SMSE}
\alias{salmonMSE-int}
\alias{SOM2MOM}
\title{Internal salmonMSE functions for converting operating model inputs and outputs}
\usage{
MMSE2SMSE(MMSE, SOM, Harvest_MMP, N, Ford)

SOM2MOM(SOM)
}
\arguments{
\item{MMSE}{Object of class \linkS4class{MMSE} returned from MSEtool}

\item{SOM}{An object of class \linkS4class{SOM}}

\item{Harvest_MMP}{Optional harvest function created by \code{\link[=make_Harvest_MMP]{make_Harvest_MMP()}}}

\item{N}{Data frame of state variables saved in the \link{salmonMSE_env} environment during the simulation}

\item{Ford}{Data frame of phenotypic trait values saved in the \link{salmonMSE_env} environment during the simulation}
}
\value{
\code{MMSE2SMSE}: \linkS4class{SMSE} object

\code{SOM2MOM}: \linkS4class{MOM} object
}
\description{
\itemize{
\item \code{\link[=SOM2MOM]{SOM2MOM()}} converts a salmon operating model (\linkS4class{SOM}) to a multi-stock operating model (\linkS4class{MOM})
\item \code{\link[=MMSE2SMSE]{MMSE2SMSE()}} converts the openMSE output, along with additional state variables recorded in \link{salmonMSE_env}, into a salmon MSE object (\linkS4class{SMSE})
\code{\link[=salmonMSE]{salmonMSE()}} is the wrapper function that coordinates the simulation and the output.
}
}
