% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonMSE.R, R/simpleSOM.R
\name{salmonMSE}
\alias{salmonMSE}
\alias{ProjectSOM}
\alias{simple_salmonMSE}
\title{Run salmonMSE}
\usage{
salmonMSE(SOM, ncores = 1, silent = FALSE)

ProjectSOM(SOM, sims, check = FALSE)

simple_salmonMSE(simpleSOM, ...)
}
\arguments{
\item{SOM}{An object of class \linkS4class{SOM}}

\item{ncores}{Integer, maximum number of processors to run projection with parallel processing}

\item{silent}{Logical, whether to report progress in console}

\item{sims}{Optional integer vector to run projection for a subset of simulations. Intended for parallel processing.}

\item{check}{Logical, whether to check the structure of the input object with \code{\link[=check_SOM]{check_SOM()}}}

\item{simpleSOM}{An object of class \linkS4class{simpleSOM}}

\item{...}{Other arguments to pass to \code{salmonMSE()}}
}
\value{
\linkS4class{SMSE} object
}
\description{
\code{salmonMSE()} runs a salmon management strategy evaluation from an operating model object (\linkS4class{SOM}), by checking
the operating model object with \code{check_SOM()}, running the projection in \code{ProjectSOM()} (parallel if called upon, then stitches
together the output in a single object), and calculates reference points with \code{calc_ref()}.

\code{ProjectSOM()} is the internal projection function.

\code{simple_salmonMSE()} is helper function that converts a simple operating model (modeling only recruit-spawners
with Ricker relationship, terminal marine fishery, without hatchery production) to a full operating model and runs the projection.
}
\examples{
\dontrun{
SMSE <- salmonMSE(simple_SOM)
}

}
