% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonMSE_env.R
\docType{data}
\name{salmonMSE_env}
\alias{salmonMSE_env}
\title{Environment to store salmon specific variables, e.g, fitness, during the simulation}
\format{
An object of class \code{environment} of length 3.
}
\usage{
salmonMSE_env
}
\description{
Environment to store salmon specific variables, e.g, fitness, during the simulation
}
\keyword{datasets}
