% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_func.R
\name{stray_func}
\alias{stray_func}
\title{Internal stray function}
\usage{
stray_func(N, stray_matrix, m)
}
\arguments{
\item{N}{Array \verb{[ns, nage, n_r]} of hatchery-origin fish}

\item{stray_matrix}{Matrix \verb{[ns, ns]} that specifies proportion that move from population in the i-th row to the j-th column. The diagonal informs non-straying proportion}

\item{m}{Vector \verb{[ns]} mark rater of hatchery-origin fish by donor population}
}
\value{
Named list:
\itemize{
\item \code{N_remain} Abundance of hatchery-origin fish that do not stray, same dimension as \code{N}
\item \code{N_stray} Abundance of strays by recipient population, same dimension as \code{N}
\item \code{m_stray} Mark rate of strays by recipient population, vector \verb{[ns]}
}
}
\description{
Calculates the number of fish that move from donor to recipient populations from a movement matrix
}
\keyword{internal}
