% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHA.R
\name{AHA}
\alias{AHA}
\title{Implementation of All-H Analyzer in R}
\usage{
AHA(SOM, ngen = 100)
}
\arguments{
\item{SOM}{An object of class \linkS4class{SOM}}

\item{ngen}{Integer, the number of generations for which to run the simulation}
}
\description{
Wrapper function for an implementation of All-H Analyzer (AHA) in R. Can be used to compare outputs between AHA
and salmonMSE.
}
