% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Bio-class}
\alias{Bio-class}
\title{Class \code{"Bio"}}
\description{
The component of the operating model that controls biological dynamics, i.e., natural production.
}
\details{
Various parameters can be stochastic (length \code{nsim}) or input as a single numeric
(value identical across all simulations).
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{nsim}}{Integer. Number of simulations.}

\item{\code{maxage}}{Integer. The maximum age of the population age structure.}

\item{\code{p_mature}}{Vector of length \code{maxage}. The proportion mature by age.}

\item{\code{capacity_smolt}}{Vector length \code{nsim}. The asymptote of the Beverton-Holt stock-recruit function for
density-dependent natural smolt production from fry life stage. Units of spawners.}

\item{\code{prod_smolt}}{Vector length \code{nsim}. The productivity parameter of the Beverton-Holt stock-recruit function.
Natural per-capita smolt production as the population approaches zero (density-independent component). Units of smolts per spawner,
frequently equivalent to recruits per spawner divided by marine survival.}

\item{\code{SAR_NOS}}{Vector length \code{nsim}. Survival to adult return of natural origin fish (density-independent, time-invariant).}

\item{\code{fec}}{Numeric. Female fecundity of natural origin spawners.}

\item{\code{p_female}}{Numeric. The proportion of females in the spawning population.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Bio")}
}

\examples{
showClass("Bio")
}
\keyword{classes}
