% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Bio-class}
\alias{Bio-class}
\title{Class \code{"Bio"}}
\description{
The component of the operating model that controls biological dynamics, i.e., natural production.
}
\details{
Various parameters can be stochastic (length \code{nsim}) or input as a single numeric
(value identical across all simulations).
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{nsim}}{Integer. Number of simulations.}

\item{\code{maxage}}{Integer. The maximum age of the population age structure.}

\item{\code{p_mature}}{Either vector by age (length \code{maxage}) or an array with dimension \verb{[nsim, maxage, nyears+proyears]}. The proportion mature by age.}

\item{\code{SRrel}}{Character, stock-recruit relationship for density-dependent smolt production. Either "BH" (Beverton-Holt) or "Ricker"}

\item{\code{capacity_smolt}}{Vector length \code{nsim}. The asymptote of the Beverton-Holt stock-recruit function, or the Ricker maximum for
density-dependent natural smolt production from fry life stage. \strong{Units of smolts.}}

\item{\code{kappa}}{Vector length \code{nsim}. The compensation ratio for the stock-recruit function, a.k.a. adult productivity.
Natural per-capita production of recruits as the population approaches zero (density-independent component). \strong{Units of recruits per egg}.
In stage-based models, equivalent to the product of smolt productivity (smolts per spawner) and marine survival.}

\item{\code{Smax}}{Vector length \code{nsim}. The spawning output that maximizes smolt production in the Ricker stock-recruit function. \strong{Units of egg production.}}

\item{\code{phi}}{Optional parameter, vector length \code{nsim}. Unfished recruits per spawner.
The \code{alpha} parameter of the stock-recruit function will be the ratio of \code{kappa} and \code{phi}. In stage-based models,
the product of marine survival, fecundity, and proportion female. If not provided, \code{phi} will be calculated from \code{Mocean_NOS} corresponding
to the first year.}

\item{\code{Mocean_NOS}}{Either vector by age (length \code{maxage}) or an array with dimension \verb{[nsim, maxage, nyears+proyears]}.
Natural mortality of immature natural origin fish.
To replicate the SAR parameter of a stage-specific model, set \code{Mocean_NOS[a] = -log(SAR)} where \code{a} is the age class prior to maturation (and zero for all other ages).}

\item{\code{fec}}{Vector by age (length \code{maxage}). Female fecundity of natural origin spawners.}

\item{\code{p_female}}{Numeric. The proportion of females in the spawning population.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Bio")}
}

\examples{
showClass("Bio")
}
\keyword{classes}
