% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM2MOM_int.R
\name{calc_phi}
\alias{calc_phi}
\title{Calculate unfished egg per smolt with life history groups}
\usage{
calc_phi(Mjuv, p_mature, p_female, fec, s_enroute = 1, n_g, p_LHG)
}
\arguments{
\item{Mjuv}{Matrix \verb{[maxage, n_g]}. Juvenile natural mortality}

\item{p_mature}{Matrix \verb{[maxage, n_g]}. Maturity at age}

\item{p_female}{Numeric. Proportion female}

\item{fec}{Matrix \verb{[maxage, n_g]}. Fecundity at age}

\item{s_enroute}{Numeric, en-route survival of escapement to spawning grounds}

\item{n_g}{Integer. Number of life history groups}

\item{p_LHG}{Proportion Mjuv Array \verb{[nsim, maxage, n_g]}. Juvenile natural mortality}
}
\description{
Calculate unfished egg per smolt with life history groups
}
\keyword{internal}
