% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRR.R
\name{calc_SRR}
\alias{calc_SRR}
\title{Calculate abundance from density-dependent mortality}
\usage{
calc_SRR(N1, N2 = N1, p, capacity, type = c("BH", "Ricker"))
}
\arguments{
\item{N1}{Numeric, the initial abundance that scales the density-independent survival term}

\item{N2}{Numeric, the initial abundance that scales the density-dependent survival term}

\item{p}{Numeric, the productivity parameter that sets the maximum survival as the initial abundance approaches zero}

\item{capacity}{Numeric, the capacity parameter that set the maximum survivors}

\item{type}{Character, the functional form of the stock-recruit relationship}
}
\value{
\itemize{
\item \code{calc_SRR} calculates the abundance of survivors
\item \code{calc_SRRpars} calculates the alpha and beta terms when the productivity parameter is in terms of abundance but N1 and N2 is in terms of
}
}
\description{
Calculates the abundance of survivors after applying either a Beverton-Holt or Ricker stock-recruit relationship.
}
\details{
The Beverton-Holt stock recruit relationship is of the following form:
\deqn{\textrm{Smolt} = \dfrac{\alpha N_1}{1 + \beta N_2}}
where \eqn{\alpha = P}, \eqn{\beta = P/C}.

The Ricker stock recruit relationship is of the following form:
\deqn{\textrm{Smolt} = \alpha N_1 \exp(-\beta N_2)}
where \eqn{\alpha = P}, \eqn{\beta = P/(Ce)}, \eqn{e} is Euler's number.

Productivity \eqn{P} is in terms of abundance per unit of \eqn{N_1} and \eqn{N_2}.
}
\seealso{
\code{\link[=calc_SRRpars]{calc_SRRpars()}}
}
