% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_func.R
\name{catch_func}
\alias{catch_func}
\title{Catch function}
\usage{
catch_func(
  NO,
  HO,
  type = c("u", "catch"),
  U,
  K,
  V,
  MSF = FALSE,
  m = 1,
  release_mort = 0
)
}
\arguments{
\item{NO}{Array \verb{[ns, nage, n_g]} of natural-origin fish}

\item{HO}{Array \verb{[ns, nage, n_r]} of hatchery-origin fish}

\item{type}{Character. Whether to calculate removals from harvest rate \code{"u"}or catch \code{"catch"}}

\item{U}{Numeric or function. Harvest rate of fishery}

\item{K}{Numeric or function. Total catch of the fishery}

\item{V}{Matrix \verb{[ns, nage]} Relative vulnerability by age class to fishery}

\item{MSF}{Logical, whether fishing is mark-selective}

\item{m}{Numeric vector length \verb{[ns]}, mark rate of hatchery-origin fish. Only used if \code{MSF = TRUE}.}

\item{release_mort}{Numeric vector length \verb{[ns]}, proportion of released unmarked fish that die. Only used if \code{MSF = TRUE}.}
}
\value{
Named list:
\itemize{
\item \code{K_NO} kept natural-origin catch, same dimension as \code{NO}
\item \code{K_HO} kept hatchery-origin catch, same dimension as \code{HO}
\item \code{D_NO} discarded (live + dead) natural-origin catch, same dimension as \code{NO}
\item \code{D_HO} discarded (live + dead) hatchery-origin catch, same dimension as \code{HO}
\item \code{DD_NO} dead discarded natural-origin catch, same dimension as \code{NO}
\item \code{DD_HO} dead discarded hatchery-origin catch, same dimension as \code{HO}
\item \code{U_NO} natural-origin harvest rate (ratio of kept catch and abundance), same dimension as \code{NO}
\item \code{U_HO} hatchery-origin harvest rate (ratio of dead catch and abundance), same dimension as \code{HO}
\item \code{Ex_NO} natural-origin exploitation rate (ratio of dead catch and abundance), same dimension as \code{NO}
\item \code{Ex_HO} hatchery-origin exploitation rate (ratio of dead catch and abundance), same dimension as \code{HO}
}
}
\description{
Internal function that does catch calculations for the preterminal and terminal marine fisheries.
}
\keyword{internal}
