% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.R
\name{fit_CM}
\alias{fit_CM}
\alias{sample_CM}
\title{Fit conditioning model to historical data}
\usage{
fit_CM(
  data,
  start = list(),
  lower_b1,
  upper_b1,
  lower_b,
  upper_b,
  do_fit = TRUE,
  silent = TRUE,
  ...
)

sample_CM(fit, ...)
}
\arguments{
\item{data}{A list containing data inputs. See details.}

\item{start}{An optional list containing parameter starting values. See details.}

\item{lower_b1}{Lower bound of coefficients for linear covariates that predict natural mortality for age 1. See details.}

\item{upper_b1}{Upper bound of coefficients for linear covariates that predict natural mortality for age 1. See details.}

\item{lower_b}{Lower bound of coefficients for linear covariates that predict natural mortality for age 2+ See details.}

\item{upper_b}{Upper bound of coefficients for linear covariates that predict natural mortality for age 2+. See details.}

\item{do_fit}{Logical, whether to do the fit and estimate the Hessian.}

\item{silent}{Logical, whether to silence output from RTMB to the console.}

\item{...}{For \code{fit_CM}, arguments to \code{\link[RTMB:TMB-interface]{RTMB::MakeADFun()}}. For \code{sample_CM}, arguments to \code{rstan::sampling()}}

\item{fit}{List of output from \code{fit_CM()}}
}
\value{
\itemize{
\item \code{fit_CM()} returns a list containing the RTMB model, nlminb output, standard errors, and parameter bounds
\item \code{sample_CM()} returns a \code{stanfit} object containing the MCMC chains
}
}
\description{
Bayesian stock reconstruction model of natural and hatchery origin fish,
fitted to coded wire tag data, observed escapement, and hatchery releases.
Estimates time-varying maturity rate as well as time-varying ocean survival as a linear model of covariates (separate covariates
assumed between age 1 vs. ages 2+).
Currently models only a preterminal fishery.
\code{fit_CM()} generates the RTMB model from data which can then be passed to \code{sample_CM()} to run the MCMC in Stan.
}
\section{Data}{

Data should passed through a named list with the following:
\itemize{
\item \code{Nages} Integer, number of age classes in the model
\item \code{Ldyr} Integer, number of years in the model
\item \code{lht} Integer, life history type. Should be 1 for now
\item \code{hatchsurv} Numeric, survival of hatchery releases into the smolt life stage. Density-independent.
\item \code{ssum} Numeric, proportion of spawners that is female
\item \code{fhist} Numeric, historical fishing mortality for calculating the spawners at age in the first year of the model
\item \code{bmatt} Vector length \code{Nages}. Proportion maturity at age, base values for calculating the unfished replacement line
\item \code{fec} Vector length \code{Nages}. Fecundity, egg production at age
\item \code{vul} Vector length \code{Nages}. Vulnerability at age to the preterminal fishery
\item \code{mobase}. Vector length \code{Nages}. Natural mortality at age, base values for calculating the unfished replacement line
\item \code{cwtrelease} Vector length \code{Ldyr}, coded wire tag (CWT) releases
\item \code{cwtesc} Matrix \verb{[Ldyr, Nages]}. CWT escapement \strong{by brood year and age}. Poisson likelhood.
\item \code{cwtcat} Matrix \verb{[Ldyr, Nages]}. CWT catch, \strong{by brood year and age}. Poisson likelhood.
\item \code{RelRegF} Vector \code{Ldyr}. Trend in relative regional fishing mortality. Fishing mortality is estimated by estimating a scaling
coefficient and annual deviations from this vector.
\item \code{obsescape} Vector length \code{Ldyr}, total observed escapement (all ages and both hatchery/natural fish). Lognormal likelhood.
\item \code{propwildspawn} Vector length \code{Ldyr}, proportion of the escapement that spawn (accounts for en-route mortality and broodtake)
\item \code{hatchrelease} Vector length \code{Ldyr+1}, number of hatchery juvenile fish released
\item \code{cwtExp} Numeric, coefficient for scaling down the CWT predictions to match the observations. For example, \code{cwtExp = 10} means that
the observed values is ten times greater than the predicted value
\item \code{covariate1} Matrix \verb{Ldyr, ncov1} of linear covariates that predict natural mortality for age 1.
\item \code{covariate} Matrix \verb{Ldyr, ncov} of linear covariates that predict natural mortality for ages 2+.
\item \code{so_mu} Numeric, the prior mean for unfished spawners in logspace
\item \code{so_sd} Numeric, the prior standard deviation for unfished spawners in logspace
\item \code{so_min} Numeric, lower bound for the estimate of unfished spawners. Optional.
}
}

\section{start}{

Starting values for parameters can be provided through a named list:
\itemize{
\item \code{cr} Numeric, compensation ratio. Default is 3.
\item \code{log_so} Numeric, unfished spawners in logspace. Default is 2.
\item \code{moadd} Numeric, additive term to base natural mortality rate for age 1 juveniles. Default is zero.
\item \code{wt} Vector \code{Ldyr}. Annual deviates in natural mortality during the freshwater life stage (affects survival to smolt life stage).
Estimated with normal prior with mean zero and standard deviation \code{p$wt_sd}. Default is zero.
\item \code{wto} Vector \code{Ldyr}. Annual deviates in natural mortality for age 2+ juveniles (marine life stage).
Estimated with normal prior with mean zero and standard deviation \code{p$wto_sd}. Default is zero.
\item \code{Fbase} Numeric, scaling coefficient to estimate fishing mortality from \code{data$RelRegF}. Default is 1.
\item \code{fanomaly} Vector \code{Ldyr}. Annual deviates from \code{Fbase * data$RelRegF} to estimate fishing mortality.
Estimated with normal prior with mean zero and standard deviation \code{p$fanomaly_sd}. Default is zero.
\item \code{lnS_sd} Numeric, lognormal standard deviation of the observed escapement. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 1.
\item \code{wt_sd} Numeric, lognormal standard deviation of the age 1 (freshwater) natural mortality deviates. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 1.
\item \code{wto_sd} Numeric, lognormal standard deviation of the age 2+ (marine) natural mortality deviates. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 1.
\item \code{fanomaly_sd} Numeric, lognormal standard deviation of \code{fanomaly}. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 1.
\item \code{logit_matt} Matrix \verb{Ldyr, Nages-2} maturity at age in logit space. Age-1 maturity is fixed to zero and one at age 1 and the maximum age, respectively. Default is 0.1.
\item \code{sd_matt} Vector \code{Nages-2}. Logit standard deviation of maturity (\code{logit_matt}) by age class. Default is 0.5.
\item \code{b1} Vector \code{ncov1} of coefficients for linear covariates that predict natural mortality for age 1. Default is zero.
\item \code{b} Vector \code{ncov} of coefficients for linear covariates that predict natural mortality for ages 2+. Default is zero.
}
}

\section{Covariates on natural mortality}{


Natural mortality is modeled as the sum of a base value \eqn{M^\textrm{base}}, additional scaling factor for age 1 \eqn{M^\textrm{add}},
a linear system of covariates \eqn{X} and coefficients \eqn{b}:

\deqn{
M_{y,a} =
\begin{cases}
M^\textrm{base}_a + M^\textrm{add} + \sum_j b^1_j X^1_{y,j} & \quad a = 1
M^\textrm{base}_a + \sum_j b_j X_{y,j} & \quad a = 2, \ldots, A
\end{cases}
}
}

\seealso{
\code{\link[=CM2SOM]{CM2SOM()}}
}
\author{
Q. Huynh with Stan code provided by J. Korman and C. Walters
}
