% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.R
\name{fit_CM}
\alias{fit_CM}
\alias{sample_CM}
\title{Fit conditioning model to historical data}
\usage{
fit_CM(
  data,
  start = list(),
  lower = list(),
  upper = list(),
  do_fit = TRUE,
  silent = TRUE,
  control = list(eval.max = 1e+05, iter.max = 1e+05),
  ...
)

sample_CM(fit, ...)
}
\arguments{
\item{data}{A list containing data inputs. See details.}

\item{start}{An optional list containing parameter starting values. See details.}

\item{lower}{Named list containing lower bounds for parameters. See details.}

\item{upper}{Named list containing upper bounds for parameters. See details.}

\item{do_fit}{Logical, whether to do the fit and estimate the Hessian.}

\item{silent}{Logical, whether to silence output from RTMB to the console.}

\item{control}{List, \code{control} argument to pass to \code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{...}{For \code{fit_CM}, arguments to \code{\link[RTMB:TMB-interface]{RTMB::MakeADFun()}}. For \code{sample_CM}, arguments to \code{rstan::sampling()}}

\item{fit}{List of output from \code{fit_CM()}}
}
\value{
\itemize{
\item \code{fit_CM()} returns a named list containing the RTMB model (\code{obj}), nlminb output (\code{opt}), standard errors (\code{SD}),
and parameter bounds (\code{lower} and \code{upper})
\item \code{sample_CM()} returns a \code{stanfit} object containing the MCMC chains
}
}
\description{
Bayesian stock reconstruction model of natural and hatchery origin fish,
fitted to coded wire tag data, observed escapement, and hatchery releases (\href{https://publications.gc.ca/site/eng/9.940685/publication.html}{Korman and Walters 2024}).
Estimates time-varying maturity rate as well as time-varying ocean survival as a linear model of covariates (separate covariates
assumed between age 1 vs. ages 2+).
The model includes either a preterminal fishery, terminal fishery, or both (see Data and start sections of the documentation).
\code{fit_CM()} generates the RTMB model from data which can then be passed to \code{sample_CM()} to run the MCMC in Stan.
}
\section{Data}{

Data should passed through a named list with the following:
\itemize{
\item \code{Nages} Integer, number of age classes in the model
\item \code{Ldyr} Integer, number of years in the model
\item \code{lht} Integer, life history type. Should be 1 for now
\item \code{hatchsurv} Numeric, survival of hatchery releases into the smolt life stage. Density-independent.
\item \code{gamma} \emph{Optional}. Numeric, the relative spawning success of hatchery origin spawners. Default is 1.
\item \code{ssum} Numeric, proportion of spawners that is female
\item \code{finitPT} Numeric, initial preterminal fishing mortality for calculating the equilibrium spawners at age in the first year of the model. Default is 0.
\item \code{finitT} Numeric, initial terminal fishing mortality for calculating the equilibrium spawners at age in the first year of the model. Default is 0.
\item \code{bmatt} Vector length \code{Nages}. Proportion maturity at age, base values for calculating the unfished replacement line.
\item \code{fec} Vector length \code{Nages}. Fecundity, egg production at age
\item \code{vulPT} Vector length \code{Nages}. Vulnerability at age to the preterminal fishery
\item \code{vulT} Vector length \code{Nages}. Vulnerability at age to the terminal fishery
\item \code{mobase}. Vector length \code{Nages}. Natural mortality at age, base values for calculating the unfished replacement line and the
the equilibrium spawners at age.
\item \code{cwtrelease} Vector length \code{Ldyr}, coded wire tag (CWT) releases
\item \code{cwtesc} Matrix \verb{[Ldyr, Nages]}. CWT escapement \strong{by brood year and age}. Poisson likelhood.
\item \code{cwtcatPT} Matrix \verb{[Ldyr, Nages]}. CWT preterminal catch, \strong{by brood year and age}. Poisson likelhood. Set all values to zero to turn off
parameters related to the preterminal fishery.
\item \code{cwtcatT} Matrix \verb{[Ldyr, Nages]}. CWT terminal catch, \strong{by brood year and age}. Poisson likelhood. Set all values to zero to turn off
parameters related to the terminal fishery.
\item \code{RelRegFPT} Vector \code{Ldyr}. Trend in relative regional preterminal fishing mortality. Fishing mortality is estimated by estimating a scaling
coefficient and annual deviations from this vector.
\item \code{RelRegFT} Vector \code{Ldyr}. Trend in relative regional terminal fishing mortality.
\item \code{obsescape} Vector length \code{Ldyr}, total observed escapement (all ages and both hatchery/natural fish). Lognormal likelhood.
\item \code{propwildspawn} Vector length \code{Ldyr}, proportion of the escapement that spawn (accounts for en-route mortality and broodtake)
\item \code{hatchrelease} Vector length \code{Ldyr+1}, number of hatchery juvenile fish released
\item \code{cwtExp} Numeric, the CWT sampling rate. This coefficient scales down the CWT predictions to match the observations. For example, \code{cwtExp = 0.1}
reduces the CWT predictions by 0.1 for the likelihood. Default is 1.
\item \code{covariate1} \emph{Optional}. Matrix \verb{Ldyr, ncov1} of linear covariates that predict natural mortality for age 1.
\item \code{covariate} \emph{Optional}. Matrix \verb{Ldyr, ncov} of linear covariates that predict natural mortality for ages 2+.
\item \code{s_enroute} Numeric, survival of escapement to spawning grounds. Default is 1.
\item \code{so_mu} Numeric, the prior mean for unfished spawners in logspace. Default is \code{log(3 * max(data$obsescape))}.
\item \code{so_sd} Numeric, the prior standard deviation for unfished spawners in logspace. Default is 0.5.
\item \code{fitness} Logical, whether to calculate fitness effects on survival. Default is \code{FALSE}.
\item \code{theta} Vector length 2, the optimum phenotype value for the natural and hatchery environments. Default is 100 and 80, respectively. See
\href{https://docs.salmonmse.com/articles/equations.html#fitness-effects-on-survival}{online article} for more information.
\item \code{rel_loss} Vector length 3, the loss in fitness apportioned between the egg, fry (both prior to density-dependence), and smolt (after density-dependence) life stage. The three values should sum to 1.
\item \code{zbar_start} Vector length 2, the mean phenotype of the spawners and broodtake in the natural and hatchery environment, respectively, at the start of the model. Default values of 100 and 100, implying maximum fitness at
for the natural environment at the start of the model.
\item \code{fitness_variance} Numeric. The variance of the phenotypic trait. Assumed identical between the natural and hatchery environments. Default is 10.
\item \code{selection_strength} Numeric. The ratio between the fitness standard deviation and the phenotype standard deviation. Default is 3.
\item \code{heritability} Numeric. The heritability of the phenotypic trait. Between 0-1. Default is 0.5.
\item \code{fitness_floor} Numeric. The minimum fitness value in the natural and hatchery environments. Default is 0.5.
}
}

\section{start}{

Starting values for parameters can be provided through a named list:
\itemize{
\item \code{log_cr} Numeric, log of the compensation ratio (productivity). Default is 3.
\item \code{log_so} Numeric, unfished spawners in logspace. Default is \code{log(3 * max(data$obsescape))}.
\item \code{moadd} Numeric, additive term to base natural mortality rate for age 1 juveniles. Default is zero.
\item \code{wt} Vector \code{Ldyr}. Annual deviates in natural mortality during the freshwater life stage (affects survival to smolt life stage).
Estimated with normal prior with mean zero and standard deviation \code{p$wt_sd}. Default is zero.
\item \code{wto} Vector \code{Ldyr}. Annual deviates in natural mortality for age 2+ juveniles (marine life stage).
Estimated with normal prior with mean zero and standard deviation \code{p$wto_sd}. Default is zero.
\item \code{FbasePT} Numeric, scaling coefficient to estimate preterminal fishing mortality from \code{data$RelRegFPT}. Default is 1.
\item \code{FbaseT} Numeric, scaling coefficient to estimate preterminal fishing mortality from \code{data$RelRegFT}. Default is 1.
\item \code{fanomalyPT} Vector \code{Ldyr}. Annual deviates from \code{FbasePT * data$RelRegFPT} to estimate preterminal fishing mortality.
Estimated with normal prior with mean zero and standard deviation \code{p$fanomaly_sd}. Default is zero.
\item \code{fanomalyT} Vector \code{Ldyr}. Annual deviates from \code{FbaseT * data$RelRegFT} to estimate terminal fishing mortality.
Estimated with normal prior with mean zero and standard deviation \code{p$fanomalyPT_sd}. Default is zero.
\item \code{lnE_sd} Numeric, lognormal standard deviation of the observed escapement. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 0.1.
\item \code{wt_sd} Numeric, lognormal standard deviation of the age 1 (freshwater) natural mortality deviates. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 1.
\item \code{wto_sd} Numeric, lognormal standard deviation of the age 2+ (marine) natural mortality deviates. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 1.
\item \code{fanomalyPT_sd} Numeric, lognormal standard deviation of \code{fanomalyPT}. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 1.
\item \code{fanomalyT_sd} Numeric, lognormal standard deviation of \code{fanomalyT}. Estimated with hierarchical \code{gamma(2, 5)} prior. Default is 1.
\item \code{logit_matt} Matrix \verb{Ldyr, Nages-2} maturity by year and age in logit space. Maturity is fixed to zero and one at age 1 and the maximum age, respectively. Default is 0.1.
\item \code{sd_matt} Vector \code{Nages-2}. Logit standard deviation of maturity (\code{logit_matt}) by age class. Default is 0.5.
\item \code{b1} Vector \code{ncov1} of coefficients for linear covariates that predict natural mortality for age 1. Default is zero.
\item \code{b} Vector \code{ncov} of coefficients for linear covariates that predict natural mortality for ages 2+. Default is zero.
}
}

\section{Bounds}{


By default, the standard deviation parameters and parameters in normal space (e.g., \code{FbasePT}, \code{Fbase_T}) have a lower bound of zero.
\code{moadd} has a lower bound of zero by default, but it is feasible that this parameter can be negative as well.
Deviation parameters centred around zero are bounded between -3 to 3.
The \code{log_cr} parameter has a lower bound of zero.

All other parameters are unbounded.
}

\section{Covariates on natural mortality}{


Natural mortality is modeled as the sum of a base value \eqn{M^\textrm{base}}, additional scaling factor for age 1 \eqn{M^\textrm{add}},
a linear system of covariates \eqn{X} and coefficients \eqn{b}:

\deqn{
M_{y,a} =
\begin{cases}
M^\textrm{base}_a + M^\textrm{add} + \sum_j b^1_j X^1_{y,j} & \quad a = 1\\
M^\textrm{base}_a + \sum_j b_j X_{y,j} & \quad a = 2, \ldots, A
\end{cases}
}
}

\references{
Korman, J. and Walters, C. 2024. A life cycle model for Chinook salmon population dynamics. Canadian Contractor Report of Hydrography and Ocean
Sciences 62: vi + 60 p.
}
\seealso{
\code{\link[=CM2SOM]{CM2SOM()}}
}
\author{
Q. Huynh with Stan code provided by J. Korman and C. Walters
}
