% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Habitat-class}
\alias{Habitat-class}
\title{Class \code{"Habitat"}}
\description{
The component of the operating model that controls habitat management.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{capacity_smolt_improve}}{Numeric. The proportional change in the smolt capacity parameter in the projection period.
For example, 1.1 represents a 10 percent improvement.}

\item{\code{prod_smolt_improve}}{Numeric. The proportional change in the smolt productivity parameter in the projection period.
For example, 1.1 represents a 10 percent improvement.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Habitat")}
}

\examples{
showClass("Habitat")
}
\keyword{classes}
