% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Harvest-class}
\alias{Harvest-class}
\title{Class \code{"Harvest"}}
\description{
The component of the operating model that controls marine harvest.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{type_PT}}{Character. Whether to manage preterminal fishery catch from exploitation rate ("u") or catch target ("catch").
Default is "u".}

\item{\code{type_T}}{Character. Whether to manage terminal fishery catch from exploitation rate ("u") or catch target ("catch").
Default is "u".}

\item{\code{u_preterminal}}{Numeric, matrix \verb{[nsim, proyears]}, or function. If \code{type_PT = "u"}, the harvest rate of the immature
component of the population in the pre-terminal fishery. The harvest rate is the ratio to kept AEQ catch to
(kept AEQ catch + return), where AEQ are adult equivalents. Function should be of the form \code{function(NO, HO, m) return(u)}.
\emph{Not used if \code{SOM@UPT_complex} is specified.}}

\item{\code{u_terminal}}{Numeric, matrix \verb{[nsim, proyears]}, or function. If \code{type_T = "u"}, the harvest rate (ratio of kept catch to
of the terminal marine fishery. Function should be of the form \code{function(NO, HO, m) return(u)}.
\emph{Not used if \code{SOM@UT_complex} is specified.}}

\item{\code{K_PT}}{Numeric or function. If \code{type_PT = "catch"}, the catch target of the immature component of the population in the
pre-terminal fishery. Function should be of the form \code{function(NO, HO, m) return(K)}.
\emph{Not used if \code{SOM@UPT_complex} is specified.}}

\item{\code{K_T}}{Numeric or function. If \code{type_T = "catch"}, the catch target of the return in the terminal fishery.
Function should be of the form \code{function(NO, HO, m) return(K)}.
\emph{Not used if \code{SOM@UT_complex} is specified.}}

\item{\code{MSF_PT}}{Logical. Whether to implement mark-selective fishing in the preterminal fishery, with no retention on unmarked fish.}

\item{\code{MSF_T}}{Logical. Whether to implement mark-selective fishing in the terminal fishery, with no retention on unmarked fish.}

\item{\code{release_mort}}{Vector length 2. The proportion of released fish that die after release, in the pre-terminal and terminal fishery.
Implemented to model mark-selective fishing. Not used if either \code{MSF_PT} or \code{MSF_T} is \code{ FALSE}.}

\item{\code{vulPT}}{Vector length \code{maxage} or matrix \verb{[nsim, maxage]}. Vulnerability schedule (between 0-1) in the preterminal fishery. Values indicate
the proportion of fishing intensity experienced by each age class.}

\item{\code{vulT}}{Vector length \code{maxage} or matrix \verb{[nsim, maxage]}. Vulnerability schedule (between 0-1) in the terminal fishery. Values indicate
the proportion of fishing intensity experienced by each age class.}

\item{\code{ForeErr_PT}}{Numeric or matrix \verb{[nsim, proyears]}. Multiplicative forecast error of the juvenile abundance in the preterminal fishery, i.e., observation error.
Only used if \code{u_preterminal} or \code{K_PT} is a function where fishing intensity is determined by abundance. Default is 1 (no error).}

\item{\code{ForeErr_T}}{Numeric or matrix \verb{[nsim, proyears]}. Multiplicative forecast error of the return size for the terminal fishery, i.e., observation error.
Only used if \code{u_terminal} or \code{K_T} is a function where fishing intensity is determined by return size. Default is 1 (no error).}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Harvest")}
}

\examples{
showClass("Harvest")
}
\keyword{classes}
