% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Harvest-class}
\alias{Harvest-class}
\title{Class \code{"Harvest"}}
\description{
The component of the operating model that controls harvest.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{u_preterminal}}{Numeric. The exploitation rate of the immature stock in the pre-terminal fishery.}

\item{\code{u_terminal}}{Numeric. The exploitation rate of the return in the terminal fishery.}

\item{\code{m}}{Numeric. The mark rate of hatchery origin fish, which affects fishery selectivity. Mark-selective fishing (MSF) is implemented if \code{m > 0}.}

\item{\code{release_mort}}{Numeric. The proportion of released fish that die due to release. Implemented to model mark-selective fishing. Not used if \code{m = 0}.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Harvest")}
}

\examples{
showClass("Harvest")
}
\keyword{classes}
