% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Harvest-class}
\alias{Harvest-class}
\title{Class \code{"Harvest"}}
\description{
The component of the operating model that controls harvest.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{u_preterminal}}{Numeric. The exploitation rate of the immature stock in the pre-terminal fishery.}

\item{\code{u_terminal}}{Numeric. The exploitation rate of the return in the terminal fishery.}

\item{\code{MSF}}{Logical. Whether to implement mark-selective fishing, with no retention on unmarked fish.}

\item{\code{release_mort}}{Vector length 2. The proportion of released fish that die after release, in the pre-terminal and terminal fishery.
Implemented to model mark-selective fishing. Not used if \code{MSF = FALSE}.}

\item{\code{vulPT}}{Vector length \code{maxage} or matrix \verb{[nsim, maxage]}. Vulnerability schedule (between 0-1) in the preterminal fishery.}

\item{\code{vulT}}{Vector length \code{maxage} or matrix \verb{[nsim, maxage]}. Vulnerability schedule (between 0-1) in the terminal fishery.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Harvest")}
}

\examples{
showClass("Harvest")
}
\keyword{classes}
