% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Hatchery-class}
\alias{Hatchery-class}
\title{Class \code{"Hatchery"}}
\description{
The component of the operating model that controls the hatchery management.
}
\details{
Various parameters can be stochastic (length \code{nsim}) or input as a single numeric
(value identical across all simulations).
}
\section{Slots}{

\describe{
\item{\code{Name}}{Identifying name.}

\item{\code{n_yearling}}{Numeric. The number of yearlings released. No hatchery is modeled if \code{n_yearling} and
\code{n_subyearling} are zero.}

\item{\code{n_subyearling}}{Numeric. The number of subyearlings released. No hatchery is modeled if \code{n_yearling} and
\code{n_subyearling} are zero.}

\item{\code{s_prespawn}}{Vector length \code{nsim}. The survival of broodtake prior to egg production. Default is 1.}

\item{\code{s_egg_smolt}}{Vector length \code{nsim}. The survival of eggs to the smolt life stage (for yearling release).}

\item{\code{s_egg_subyearling}}{Vector length \code{nsim}. The survival of eggs to subyearling life stage (for subyearling release).}

\item{\code{gamma}}{Vector length \code{nsim}. The relative fecundity of hatchery origin spawners relative to natural origin spawners.}

\item{\code{pmax_NOB}}{Numeric. The maximum ratio of natural origin broodtake to the natural origin escapement.}

\item{\code{ptarget_NOB}}{Numeric. The target proportion of the natural origin broodtake relative to the overall broodtake.}

\item{\code{premove_HOS}}{Numeric. The proportion of the hatchery origin spawners removed from the spawning grounds (in order to
ensure a high proportion of NOS).}

\item{\code{fec_brood}}{Numeric. The fecundity of broodtake.}

\item{\code{fitness_type}}{Character. Fitness function for hatchery origin spawners. Either "Ford", "Busack", or "none".}

\item{\code{theta}}{Vector length 3.}

\item{\code{rel_loss}}{Vector length 3.}

\item{\code{Zpop_start}}{Vector length 3.}

\item{\code{fitness_variance}}{Numeric.}

\item{\code{selection_strength}}{Numeric.}

\item{\code{heritability}}{Numeric.}

\item{\code{fitness_floor}}{Numeric.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Hatchery")}
}

\examples{
showClass("Hatchery")
}
\keyword{classes}
