% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Historical-class}
\alias{Historical-class}
\title{Class \code{"Historical"}}
\description{
Optional component of the operating model that specifies the historical dynamics. The default option starts the projection with 1000
natural-origin juveniles and 1000 hatchery-origin juveniles (if there is hatchery production) in the oldest age class, representing
single-brood year returns since there is only one age class in the population. Specify the abundance in all age class to simulate multiple
brood-year returns.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{InitNjuv_NOS}}{Array by \verb{[nsim, maxage, n_g]}. The abundance of immature natural origin fish at the beginning of the projection.
Default assumes 1000 in the oldest age class, which creates a population with single brood year returns.}

\item{\code{InitNjuv_HOS}}{Array by \verb{[nsim, maxage, n_r]}. The abundance of immature hatchery origin fish at the beginning of the projection.
Default assumes 1000 in the oldest age class, which creates a population with single brood year returns.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Historical")}
}

\examples{
showClass("Historical")
}
\keyword{classes}
