% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{Historical-class}
\alias{Historical-class}
\title{Class \code{"Historical"}}
\description{
The component of the operating model that specifies the historical dynamics.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{HistSmolt}}{Vector by historical years (\code{nyears}) or a matrix by \verb{[nsim, nyears]}. The natural smolt production.}

\item{\code{HistSpawner}}{Array by \verb{[nsim, maxage, nyears]}. Total spawners (natural and hatchery origin combined) in the natural environment.}

\item{\code{HistN}}{Array by \verb{[nsim, maxage, nyears, 2]}. The abundance of immature fish at the beginning of the annual time step. The first array slice
corresponds to natural origin fish and the second array slice corresponds to hatchery origin fish.}

\item{\code{HistYearling}}{Vector by historical years. Hatchery releases (number of smolts).}

\item{\code{HistFPT}}{Vector by historical years (\code{nyears}) or an array by dimension \verb{[nsim, nyears, 2]}. The instantaneous fishing mortality in the preterminal fishery.
The first array slice corresponds to F for natural origin fish and the second array slice corresponds to hatchery origin fish.}

\item{\code{HistFT}}{Vector by historical years (\code{nyears}) or an array by dimension \verb{[nsim, nyears, 2]}. The instantaneous fishing mortality in the terminal fishery.
The first array slice corresponds to F for natural origin fish and the second array slice corresponds to hatchery origin fish.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("Historical")}
}

\examples{
showClass("Historical")
}
\keyword{classes}
