% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-LHG.R
\name{plot_LHG}
\alias{plot_LHG}
\title{Plot life history groups}
\usage{
plot_LHG(
  SMSE,
  var = "NOS",
  s = 1,
  FUN = median,
  figure = TRUE,
  xlab = "Projection Year",
  ylab = paste("Proportion", var),
  LHG_names
)
}
\arguments{
\item{SMSE}{Class \linkS4class{SMSE} object returned by \code{\link[=salmonMSE]{salmonMSE()}}}

\item{var}{Character. Slot for the state variables in \code{SMSE@Misc$LHG[[1]]}.}

\item{s}{Integer. Population index for multi-population model (e.g., \code{s = 1} is the first population in the model)}

\item{FUN}{Summarizing function across simulations, typically \code{\link[=median]{median()}} or \code{\link[=mean]{mean()}}}

\item{figure}{Logical, whether to generate a figure (set to FALSE if only using the function to return the data matrix)}

\item{xlab}{Character. Name of time variable for the figure}

\item{ylab}{Character. Name of the state variable for the figure}

\item{LHG_names}{Character. Vector of names for the life history groups}
}
\description{
Plot the annual proportions of life history groups at various life stages
}
\seealso{
\code{\link[=plot_statevar_ts]{plot_statevar_ts()}}
}
