% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statevar.R
\name{plot_statevar_ts}
\alias{plot_statevar_ts}
\alias{plot_statevar_hist}
\title{Plot output from salmonMSE}
\usage{
plot_statevar_ts(
  SMSE,
  var = "PNI",
  s = 1,
  figure = TRUE,
  xlab = "Projection Year",
  quant = FALSE,
  ylab = var,
  ylim,
  ...
)

plot_statevar_hist(SMSE, var = "PNI", s = 1, y, figure = TRUE, xlab = var, ...)
}
\arguments{
\item{SMSE}{Class \linkS4class{SMSE} object returned by \code{\link[=salmonMSE]{salmonMSE()}}}

\item{var}{Character. Slot for the state variable in \code{SMSE} object. See \code{slotNames(SMSE)} for options.}

\item{s}{Integer. Population index for multi-population model (e.g., \code{s = 1} is the first population in the model)}

\item{figure}{Logical, whether to generate a figure (set to FALSE if only using the function to return the data matrix)}

\item{xlab}{Character. Name of time variable for the figure}

\item{quant}{Logical, whether to plot individual simulations (FALSE) or the median with 95 percent confidence intervals (TRUE)}

\item{ylab}{Character. Name of the state variable for the figure}

\item{ylim}{Vector. Y-axis limits}

\item{...}{Additional arguments to base plot function}

\item{y}{Integer. Projection year for the state variable to plot the histogram}
}
\value{
Functions return the matrix of plotted values invisibly. Figure plotted from base graphics
}
\description{
Various functions that plot the state variables from salmonMSE projections:
\itemize{
\item \code{plot_statevar_ts()} produces a time series for all simulations, or with confidence intervals
\item \code{plot_statevar_hist()} produces a histogram across all simulations for a particular year
}
}
\seealso{
\code{\link[=plot_decision_table]{plot_decision_table()}}
}
