% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE2SMSE.R, R/SOM2MOM.R, R/SOM2MOM_int.R
\name{MMSE2SMSE}
\alias{MMSE2SMSE}
\alias{salmonMSE-int}
\alias{SOM2MOM}
\alias{make_Stock}
\alias{make_Fleet}
\title{Internal salmonMSE functions for converting operating model inputs and outputs}
\usage{
MMSE2SMSE(MMSE, SOM, Harvest_MMP, N, Ford, state)

SOM2MOM(SOM, start = list())

make_Stock(
  SOM,
  NOS = TRUE,
  stage = c("immature", "return", "escapement"),
  start = list()
)

make_Fleet(SOM, NOS = TRUE, stage = c("immature", "return", "escapement"))
}
\arguments{
\item{MMSE}{Object of class \linkS4class{MMSE} returned from MSEtool}

\item{SOM}{An object of class \linkS4class{SOM}}

\item{Harvest_MMP}{Optional harvest function created by \code{\link[=make_Harvest_MMP]{make_Harvest_MMP()}}}

\item{N}{Data frame of abundance at age saved in the \link{salmonMSE_env} environment during the simulation}

\item{Ford}{Data frame of phenotypic trait values saved in the \link{salmonMSE_env} environment during the simulation}

\item{state}{Data frame of state variables saved in the \link{salmonMSE_env} environment during the simulation}

\item{start}{An optional named list to specify the natural smolt production ('Natural') and smolt releases ('Hatchery') at the start of the
simulation}

\item{NOS}{Logical, whether the Stock or Fleet object corresponds to natural origin or hatchery origin fish}

\item{stage}{Character indicating the corresponding salmon life stage of the Stock or Fleet object}
}
\value{
\code{MMSE2SMSE}: \linkS4class{SMSE} object

\code{SOM2MOM}: \linkS4class{MOM} object

\code{make_Stock}: List containing a \linkS4class{Stock} object and accompanying custom parameters list

\code{make_Stock}: List containing a \linkS4class{Fleet} object and accompanying custom parameters list
}
\description{
\itemize{
\item \code{\link[=SOM2MOM]{SOM2MOM()}} converts a salmon operating model (\linkS4class{SOM}) to a multi-stock operating model (\linkS4class{MOM})
\item \code{\link[=MMSE2SMSE]{MMSE2SMSE()}} converts the openMSE output, along with additional state variables recorded in \link{salmonMSE_env}, into a salmon MSE object (\linkS4class{SMSE})
\item \code{\link[=make_Stock]{make_Stock()}} creates the \linkS4class{Stock} object (openMSE) corresponding to salmon life stage
\item \code{\link[=make_Fleet]{make_Fleet()}} creates the \linkS4class{Fleet} object (openMSE) corresponding to the fishery that interacts with the various salmon life stages
}

\code{\link[=salmonMSE]{salmonMSE()}} is the wrapper function that coordinates the simulation and the output.
}
\examples{
\dontrun{
# One thousand natural and hatchery smolts in the first year
SOM <- SOM2MOM(SOM, start = list(Natural = 1e3, Hatchery = 1e3))
}
}
