% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonMSE.R
\name{salmonMSE}
\alias{salmonMSE}
\alias{ProjectSOM}
\title{Run salmonMSE}
\usage{
salmonMSE(SOM, silent = FALSE)

ProjectSOM(SOM, check = FALSE)
}
\arguments{
\item{SOM}{An object of class \linkS4class{SOM}}

\item{silent}{Logical, whether to report progress in console}

\item{check}{Logical, whether to check the structure of the input object with \code{\link[=check_SOM]{check_SOM()}}}
}
\value{
\linkS4class{SMSE} object
}
\description{
\code{salmonMSE()} runs a salmon management strategy evaluation from an operating model object (\linkS4class{SOM}).

\code{ProjectSOM()} is the internal projection function.
}
\examples{
\dontrun{
SMSE <- salmonMSE(simple_SOM)
}

}
