% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM2MOM.R, R/salmonMSE.R
\name{SOM2MOM}
\alias{SOM2MOM}
\alias{salmonMSE}
\title{Run salmonMSE}
\usage{
SOM2MOM(SOM)

salmonMSE(SOM, Hist = FALSE, silent = FALSE, trace = FALSE)
}
\arguments{
\item{SOM}{An object of class \linkS4class{SOM}}

\item{Hist}{Logical, whether to stop the function stop after historical simulations? Returns a list containing all historical data}

\item{silent}{Logical, whether to report progress in console bar}

\item{trace}{Logical, whether to report additional messages from openMSE}
}
\description{
Runs a salmon management strategy evaluation through the following steps:
\itemize{
\item Converts a salmon operating model to a multi-stock operating model (MOM) via \code{SOM2MOM}
\item Creates a harvest management procedure specifying harvest rate
\item Generates the historical reconstruction of the state dynamics
\item Runs projection (if \code{Hist = TRUE})
}
}
