% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonMSE.R
\name{salmonMSE}
\alias{salmonMSE}
\title{Run salmonMSE}
\usage{
salmonMSE(
  SOM,
  start = list(),
  Hist = FALSE,
  silent = FALSE,
  trace = FALSE,
  convert = TRUE
)
}
\arguments{
\item{SOM}{An object of class \linkS4class{SOM}}

\item{start}{An optional named list to specify the natural smolt production ('Natural') and smolt releases ('Hatchery') at the start of the
simulation}

\item{Hist}{Logical, whether to stop the function stop after historical simulations? Returns a list containing all historical data}

\item{silent}{Logical, whether to report progress in console bar}

\item{trace}{Logical, whether to report additional messages from openMSE}

\item{convert}{Logical, whether to convert the output into a salmon MSE (SMSE) object}
}
\value{
If \code{Hist = TRUE} a multiHist object (list). Otherwise, if \code{convert = TRUE}, a \linkS4class{SMSE} object or if \code{convert = FALSE}, a \linkS4class{MMSE} object.
}
\description{
\code{salmonMSE()} runs a salmon management strategy evaluation through the following steps:
\itemize{
\item Converts a salmon operating model (\linkS4class{SOM}) to a multi-stock operating model (\linkS4class{MOM}) via \code{\link[=SOM2MOM]{SOM2MOM()}}
\item Creates a harvest management procedure specifying the harvest control rule
\item Generates the historical reconstruction of the state variables
\item Runs projection (if \code{Hist = FALSE})
\item Converts the openMSE output, along with additional state variables recorded in \link{salmonMSE_env}, into a salmon MSE object (SMSE) via \code{\link[=MMSE2SMSE]{MMSE2SMSE()}}
}
}
\examples{
\dontrun{
# One thousand natural and hatchery smolts in the first year
myMSE <- salmonMSE(SOM, start = list(Natural = 1e3, Hatchery = 1e3))
}
}
