% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSOM.R
\docType{class}
\name{simpleSOM-class}
\alias{simpleSOM-class}
\alias{simpleSOM}
\title{Class \code{"simpleSOM"}}
\description{
A simple operating model for modeling recruit-spawners and a terminal marine fishery, no hatchery production.
}
\details{
Various parameters can be stochastic (length \code{nsim}) or input as a single numeric
(value identical across all simulations).
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{nsim}}{Integer. Number of simulations.}

\item{\code{maxage}}{Integer. The maximum age of the population age structure.}

\item{\code{ngen}}{Integer. The number of generations (life cycles) to run the projection.}

\item{\code{kappa}}{Vector length \code{nsim}. The adult productivity ratio for the stock-recruit function. \strong{Units of recruits per spawner.}
Natural per-capita production of recruits as the population approaches zero (density-independent component).}

\item{\code{Smax}}{Vector length \code{nsim}. The spawner abundance that maximizes smolt production in the Ricker stock-recruit function. \strong{Units of spawners.}}

\item{\code{sigmaR}}{Vector length \code{nsim}. The lognormal standard deviation in the Ricker stock-recruit relationship. Recruitment anomalies
will be simulated with \verb{[stats::rlnorm()]}.}

\item{\code{Recdev}}{Optional matrix \verb{[nsim, ngen-1]} of recruitment anomalies in the Ricker stock-recruit relationship. Overrides the \code{sigmaR} argument
and provides more user flexibility, for example, simulating with auto-correlation.}

\item{\code{type_T}}{Character. Whether to manage terminal fishery catch from exploitation rate ("u") or catch target ("catch").
Default is "u".}

\item{\code{u_terminal}}{Numeric, matrix \verb{[nsim, proyears]}, or function. If \code{type_T = "u"}, the harvest rate (ratio of kept catch to
of the terminal marine fishery. Function should be of the form \code{function(NO, HO, m) return(u)}.}

\item{\code{K_T}}{Numeric or function. If \code{type_T = "catch"}, the catch target of the return in the terminal fishery.
Function should be of the form \code{function(NO, HO, m) return(K)}.}

\item{\code{InitEsc}}{Single numeric or vector \verb{[nsim]}. The escapement at the beginning of the projection. Default is 1000.}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("simpleSOM")}
}

\examples{
showClass("simpleSOM")
}
\keyword{classes}
