% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{SMSE-class}
\alias{SMSE-class}
\title{Class \code{"SMSE"}}
\description{
Stores the outputs from the simulation of salmon operating models.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{nyears}}{Integer. The number of historical years}

\item{\code{proyears}}{Integer. The number of projected years}

\item{\code{nsim}}{Integer. The number of simulations}

\item{\code{nstocks}}{Integer. The number of stocks}

\item{\code{Snames}}{Character. Stock names}

\item{\code{Fry_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Spawning output of natural origin spawners.}

\item{\code{Fry_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Spawning output of hatchery origin spawners.}

\item{\code{Smolts_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Smolts that are offspring of natural origin spawners from the previous generation.}

\item{\code{Smolts_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Smolts that are offspring of hatchery origin spawners from the previous generation.}

\item{\code{Smolts_Rel}}{Array \verb{[nsim, nstocks, proyears]}. Smolts that are offspring of broodtake from the previous generation, i.e., hatchery releases.}

\item{\code{Return_NOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. Mature fish that will be natural origin spawners.}

\item{\code{Return_HOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. Mature fish that will be hatchery origin spawners.}

\item{\code{Escapement_NOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. The escapement of mature fish that will be natural origin spawners.}

\item{\code{Escapement_HOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. The escapement of mature fish that will be hatchery origin spawners.}

\item{\code{NOB}}{Array \verb{[nsim, nstocks, nage, proyears]}. The broodtake of fish that would be natural origin spawners.}

\item{\code{HOB}}{Array \verb{[nsim, nstocks, nage, proyears]}. The broodtake of fish that would be hatchery origin spawners.}

\item{\code{NOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. Natural origin spawners.}

\item{\code{HOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. Hatchery origin spawners.}

\item{\code{HOS_effective}}{Array \verb{[nsim, nstocks, nage, proyears]}. Hatchery origin spawners discounted by \code{gamma}.}

\item{\code{Catch_NOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. Catch of natural origin spawners.}

\item{\code{Catch_HOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. Catch of hatchery origin spawners.}

\item{\code{U_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Harvest rate of natural origin spawners.}

\item{\code{U_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Harvest rate of hatchery origin spawners.}

\item{\code{fitness}}{Array \verb{[nsim, nstocks, proyears]}. Fitness.}

\item{\code{SAR_loss}}{Array \verb{[nsim, nstocks, proyears]}. Realized SAR due to fitness loss.}

\item{\code{Misc}}{List. Miscellaneous output}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("SMSE")}
}

\examples{
showClass("SMSE")
}
\keyword{classes}
