% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{SMSE-class}
\alias{SMSE-class}
\title{Class \code{"SMSE"}}
\description{
Stores the outputs from the simulation of salmon operating models.
}
\details{
In generation \eqn{t}, proportionate natural influence (PNI) is defined as:

\deqn{\textrm{PNI}_t = \dfrac{p^{\textrm{NOB}}_t}{p^{\textrm{NOB}}_t + p^{\textrm{HOS}}_t}}

The proportion of wild salmon is defined as:

\deqn{p^{\textrm{WILD}}_t = q^{\textrm{HOS}}_t
\dfrac{(q^{\textrm{HOS}}_{t-1})^2}
{(q^{\textrm{HOS}}_{t-1})^2 + 2\gamma \times p^{\textrm{HOS}}_{t-1} q^{\textrm{HOS}}_{t-1} +
\gamma^2 (p^{\textrm{HOS}}_{t-1})^2}}

where \eqn{q = 1-p}.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{nyears}}{Integer. The number of historical years}

\item{\code{proyears}}{Integer. The number of projected years}

\item{\code{nsim}}{Integer. The number of simulations}

\item{\code{nstocks}}{Integer. The number of stocks}

\item{\code{Snames}}{Character. Stock names}

\item{\code{Fry_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Spawning output of natural origin spawners.}

\item{\code{Fry_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Spawning output of hatchery origin spawners.}

\item{\code{Smolt_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Smolts that are offspring of natural origin spawners.}

\item{\code{Smolt_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Smolts that are offspring of hatchery origin spawners.}

\item{\code{Smolt_Rel}}{Array \verb{[nsim, nstocks, proyears]}. Smolts that are offspring of broodtake, i.e., hatchery releases.}

\item{\code{Return_NOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. Mature fish that will be natural origin spawners.}

\item{\code{Return_HOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. Mature fish that will be hatchery origin spawners.}

\item{\code{Escapement_NOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. The escapement of mature fish that will be natural origin spawners.}

\item{\code{Escapement_HOS}}{Array \verb{[nsim, nstocks, nage, proyears]}. The escapement of mature fish that will be hatchery origin spawners.}

\item{\code{NOB}}{Array \verb{[nsim, nstocks, proyears]}. The broodtake of natural origin spawners.}

\item{\code{HOB}}{Array \verb{[nsim, nstocks, proyears]}. The broodtake of hatchery origin spawners.}

\item{\code{NOS}}{Array \verb{[nsim, nstocks, proyears]}. Natural origin spawners.}

\item{\code{HOS}}{Array \verb{[nsim, nstocks, proyears]}. Hatchery origin spawners.}

\item{\code{HOS_effective}}{Array \verb{[nsim, nstocks, proyears]}. Hatchery origin spawners discounted by \code{gamma}.}

\item{\code{CatchPT_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Pre-terminal fishery catch of natural origin spawners.}

\item{\code{CatchT_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Terminal fishery catch of natural origin spawners.}

\item{\code{CatchPT_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Pre-terminal fishery catch of hatchery origin spawners.}

\item{\code{CatchT_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Terminal fishery catch of hatchery origin spawners.}

\item{\code{UPT_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Pre-terminal fishery harvest rate of natural origin spawners.}

\item{\code{UT_NOS}}{Array \verb{[nsim, nstocks, proyears]}. Terminal fishery harvest rate of natural origin spawners.}

\item{\code{UPT_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Pre-terminal fishery harvest rate of hatchery origin spawners.}

\item{\code{UT_HOS}}{Array \verb{[nsim, nstocks, proyears]}. Terminal fishery harvest rate of hatchery origin spawners.}

\item{\code{fitness}}{Array \verb{[nsim, nstocks, proyears]}. Fitness of the natural spawning population.}

\item{\code{PNI}}{Array \verb{[nsim, nstocks, proyears]}. Proportionate natural influence, index of gene flow from hatchery to the natural environment.}

\item{\code{p_wild}}{Array \verb{[nsim, nstocks, proyears]}. Proportion of wild spawners, whose parents also spawned in the natural environment assuming
non-assortative mating, defined under Canada's Wild Salmon Policy.}

\item{\code{SAR_loss}}{Array \verb{[nsim, nstocks, proyears]}. Realized SAR due to fitness loss.}

\item{\code{Misc}}{List. Miscellaneous output}
}}

\section{Creating Object}{

Objects can be created by calls of the form \code{new("SMSE")}
}

\examples{
showClass("SMSE")
}
\references{
Withler et al. 2018. Genetically Based Targets for Enhanced Contributions to Canadian Pacific Chinook Salmon Populations.
DFO Can. Sci. Advis. Sec. Res. Doc. 2018/019. xii + 88 p.
}
\keyword{classes}
