% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{SOM-class}
\alias{SOM-class}
\alias{SOM}
\title{Class \code{"SOM"}}
\description{
An object containing all the parameters for a salmon operating model (SOM).
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{nyears}}{Integer. The number of historical years}

\item{\code{proyears}}{Integer. The number of projected years}

\item{\code{seed}}{Integer. A random seed to ensure users can reproduce results exactly}

\item{\code{nsim}}{Integer. Number of simulations.}

\item{\code{maxage}}{Integer. The maximum age of the population age structure.}

\item{\code{p_mature}}{Vector of length \code{maxage}. The proportion mature by age.}

\item{\code{capacity_smolt}}{Vector length \code{nsim}. The asymptote of the Beverton-Holt stock-recruit function for
density-dependent natural smolt production from fry life stage.}

\item{\code{prod_smolt}}{Vector length \code{nsim}. The productivity parameter of the Beverton-Holt stock-recruit function.
Natural per-capita smolt production as the population approaches zero (density-independent component).}

\item{\code{SAR_NOS}}{Vector length \code{nsim}. Survival to adult return of natural origin fish (density-independent, time-invariant).}

\item{\code{fec}}{Numeric. Female fecundity of natural origin spawners.}

\item{\code{p_female}}{Numeric. The proportion of females in the spawning population.}

\item{\code{capacity_smolt_improve}}{Numeric. The proportional change in the smolt capacity parameter in the projection period.
For example, 1.1 represents a 10 percent improvement.}

\item{\code{prod_smolt_improve}}{Numeric. The proportional change in the smolt productivity parameter in the projection period.
For example, 1.1 represents a 10 percent improvement.}

\item{\code{n_yearling}}{Numeric. The number of yearlings released. No hatchery is modeled if both \code{n_yearling} and
\code{n_subyearling} are zero.}

\item{\code{n_subyearling}}{Numeric. The number of subyearlings released. No hatchery is modeled if both \code{n_yearling} and
\code{n_subyearling} are zero.}

\item{\code{s_prespawn}}{Numeric. The survival of broodtake prior to egg production. \code{1 - s_prespawn} is the proportion of fish not used for hatchery purposes, e.g., mortality or other resesarch purposes.
Used to back-calculate the broodtake from \code{n_yearling} and \code{n_subyearling}.}

\item{\code{s_egg_smolt}}{Numeric. The survival of eggs to the smolt life stage (for yearling release).
Used to back-calculate the broodtake from \code{n_yearling} and \code{n_subyearling}.}

\item{\code{s_egg_subyearling}}{Numeric. The survival of eggs to subyearling life stage (for subyearling release).
Used to back-calculate the broodtake from \code{n_yearling} and \code{n_subyearling}.}

\item{\code{SAR_HOS}}{Vector length \code{nsim}. Survival to adult return of hatchery origin fish (density-independent, time-invariant).}

\item{\code{gamma}}{Numeric. The relative fecundity of hatchery origin spawners relative to natural origin spawners.}

\item{\code{pmax_NOB}}{Numeric. The maximum proportion of the natural origin escapement to be used as broodtake.}

\item{\code{ptarget_NOB}}{Numeric. The target proportion of the natural origin broodtake relative to the overall broodtake.}

\item{\code{premove_HOS}}{Numeric. The proportion of the hatchery origin spawners to be removed from the spawning grounds (in order to
ensure a high proportion of NOS). This removal is also used to calculate the available hatchery origin fish for broodtake.}

\item{\code{fec_brood}}{Numeric. The fecundity of broodtake.}

\item{\code{fitness_type}}{Character. Fitness function for hatchery origin spawners. Either "Ford" or "none".}

\item{\code{theta}}{Vector length 2. The optimum phenotype value for the natural and hatchery environments.}

\item{\code{rel_loss}}{Vector length 3. The loss in fitness apportioned among the egg, fry, and smolt life stages. The three values should sum to 1.}

\item{\code{pbar_start}}{Vector length 2. The mean phenotype value in the the natural and hatchery population at the start of the projection.}

\item{\code{fitness_variance}}{Numeric. The variance of the phenotype in the population. Assumed identical between the natural and hatchery environments.}

\item{\code{selection_strength}}{Numeric. The ratio between the fitness standard deviation and the phenotype standard deviation.}

\item{\code{heritability}}{Numeric. The heritability of the phenotypic trait. Between 0-1.}

\item{\code{fitness_floor}}{Numeric. The minimum fitness value in the natural environment.}

\item{\code{u_preterminal}}{Numeric. The exploitation rate of the immature stock in the pre-terminal fishery.}

\item{\code{u_terminal}}{Numeric. The exploitation rate of the return in the terminal fishery.}

\item{\code{m}}{Numeric. The mark rate of hatchery origin fish, which affects fishery selectivity.}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("SOM", Bio, Habitat, Hatchery, Harvest)}.
}

\keyword{classes}
