% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{SOM-class}
\alias{SOM-class}
\alias{SOM}
\title{Class \code{"SOM"}}
\description{
An object containing all the parameters for a salmon operating model (SOM).
}
\section{Slots}{

\describe{
\item{\code{Name}}{Name of the operating modelx}

\item{\code{nyears}}{The number of historical years}

\item{\code{proyears}}{The number of projected years}

\item{\code{seed}}{A random seed to ensure users can reproduce results exactly}

\item{\code{Name}}{Identifying name.}

\item{\code{nsim}}{Integer. Number of simulations.}

\item{\code{maxage}}{Integer. The maximum age of the population age structure.}

\item{\code{p_mature}}{Vector of length \code{maxage}. The proportion mature by age.}

\item{\code{capacity_smolt}}{Vector length \code{nsim}. The asymptote of the Beverton-Holt stock-recruit function for
density-dependent natural smolt production from fry life stage.}

\item{\code{prod_smolt}}{Vector length \code{nsim}. The productivity parameter of the Beverton-Holt stock-recruit function.
Natural per-capita smolt production as the population approaches zero (density-independent component).}

\item{\code{SAR}}{Vector length \code{nsim}. Survival to adult return (density-independent, time-invariant).}

\item{\code{fec}}{Vector length \code{nsim}. Female fecundity of natural origin spawners.}

\item{\code{p_female}}{Vector length \code{nsim}. The proportion of females in the spawning population.}

\item{\code{Name}}{Character. Identifying name.}

\item{\code{capacity_smolt_improve}}{Numeric. The proportional change in the smolt capacity parameter in the projection period.
For example, 1.1 represents a 10 percent improvement.}

\item{\code{prod_smolt_improve}}{Numeric. The proportional change in the smolt productivity parameter in the projection period.
For example, 1.1 represents a 10 percent improvement.}

\item{\code{Name}}{Identifying name.}

\item{\code{n_yearling}}{Numeric. The number of yearlings released. No hatchery is modeled if \code{n_yearling} and
\code{n_subyearling} are zero.}

\item{\code{n_subyearling}}{Numeric. The number of subyearlings released. No hatchery is modeled if \code{n_yearling} and
\code{n_subyearling} are zero.}

\item{\code{s_prespawn}}{Vector length \code{nsim}. The survival of broodtake prior to egg production. Default is 1.}

\item{\code{s_egg_smolt}}{Vector length \code{nsim}. The survival of eggs to the smolt life stage (for yearling release).}

\item{\code{s_egg_subyearling}}{Vector length \code{nsim}. The survival of eggs to subyearling life stage (for subyearling release).}

\item{\code{gamma}}{Vector length \code{nsim}. The relative fecundity of hatchery origin spawners relative to natural origin spawners.}

\item{\code{pmax_NOB}}{Numeric. The maximum ratio of natural origin broodtake to the natural origin escapement.}

\item{\code{ptarget_NOB}}{Numeric. The target proportion of the natural origin broodtake relative to the overall broodtake.}

\item{\code{premove_HOS}}{Numeric. The proportion of the hatchery origin spawners removed from the spawning grounds (in order to
ensure a high proportion of NOS).}

\item{\code{fec_brood}}{Numeric. The fecundity of broodtake.}

\item{\code{fitness_type}}{Character. Fitness function for hatchery origin spawners. Either "Ford", "Busack", or "none".}

\item{\code{theta}}{Vector length 3.}

\item{\code{rel_loss}}{Vector length 3.}

\item{\code{Zpop_start}}{Vector length 3.}

\item{\code{fitness_variance}}{Numeric.}

\item{\code{selection_strength}}{Numeric.}

\item{\code{heritability}}{Numeric.}

\item{\code{fitness_floor}}{Numeric.}

\item{\code{Name}}{Character. Identifying name.}

\item{\code{u}}{Numeric. The exploitation rate of the return.}

\item{\code{m}}{Numeric. The mark rate of hatchery origin fish, which affects fishery selectivity.}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("SOM", Bio, Habitat, Hatchery, Harvest)}.
}

\keyword{classes}
