% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-SOM.R
\docType{class}
\name{SOM-class}
\alias{SOM-class}
\alias{SOM}
\title{Class \code{"SOM"}}
\description{
An object containing all the parameters for a salmon operating model (SOM).
}
\details{
A description of the fitness parameters and hatchery dynamics is available in the
\href{https://docs.salmonmse.com/articles/equations.html#fitness-effects-on-survival}{online documentation}.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. Identifying name}

\item{\code{nyears}}{Integer. The number of historical years}

\item{\code{proyears}}{Integer. The number of projected years}

\item{\code{seed}}{Integer. A random seed to ensure users can reproduce results exactly}

\item{\code{nsim}}{Integer. Number of simulations.}

\item{\code{maxage}}{Integer. The maximum age of the population age structure.}

\item{\code{p_mature}}{Either vector by age (length \code{maxage}) or an array with dimension \verb{[nsim, maxage, nyears+proyears]}. The proportion mature by age.}

\item{\code{SRrel}}{Character, stock-recruit relationship for density-dependent smolt production. Either "BH" (Beverton-Holt) or "Ricker"}

\item{\code{capacity_smolt}}{Vector length \code{nsim}. The asymptote of the Beverton-Holt stock-recruit function, or the Ricker maximum for
density-dependent natural smolt production from fry life stage. Units of smolts.}

\item{\code{prod_smolt}}{Vector length \code{nsim}. The productivity parameter of the Beverton-Holt stock-recruit function.
Natural per-capita smolt production as the population approaches zero (density-independent component). Units of smolts per spawner,
frequently equivalent to recruits per spawner divided by marine survival.}

\item{\code{a}}{Vector length \code{nsim}. The alpha parameter of the Ricker stock-recruit function.}

\item{\code{Smax}}{Vector length \code{nsim}. The Smax parameter of the Ricker stock-recruit function.}

\item{\code{Mocean_NOS}}{Either vector by age (length \code{maxage}) or an array with dimension \verb{[nsim, maxage, nyears+proyears]}.
Natural mortality of immature natural origin fish.
To replicate the SAR parameter of a stage-specific model, set \code{Mocean_NOS[a] = -log(SAR)} where \code{a} is the age class prior to maturation (and zero for all other ages).}

\item{\code{fec}}{Vector by age (length \code{maxage}). Female fecundity of natural origin spawners.}

\item{\code{p_female}}{Numeric. The proportion of females in the spawning population.}

\item{\code{capacity_smolt_improve}}{Numeric. The proportional change in the smolt capacity parameter in the projection period.
For example, 1.1 represents a 10 percent improvement.}

\item{\code{prod_smolt_improve}}{Numeric. The proportional change in the smolt productivity parameter in the projection period.
For example, 1.1 represents a 10 percent improvement.}

\item{\code{n_yearling}}{Numeric. The number of yearlings released. No hatchery is modeled if both \code{n_yearling} and
\code{n_subyearling} are zero.}

\item{\code{n_subyearling}}{Numeric. The number of subyearlings released. No hatchery is modeled if both \code{n_yearling} and
\code{n_subyearling} are zero.}

\item{\code{s_prespawn}}{Numeric. The survival of broodtake prior to egg production. \code{1 - s_prespawn} is the proportion of fish not used for hatchery purposes, e.g., mortality or other resesarch purposes.
Used to back-calculate the broodtake from \code{n_yearling} and \code{n_subyearling}.}

\item{\code{s_egg_smolt}}{Numeric. The survival of eggs to the smolt life stage (for yearling release).
Used to back-calculate the broodtake from \code{n_yearling} and \code{n_subyearling}.}

\item{\code{s_egg_subyearling}}{Numeric. The survival of eggs to subyearling life stage (for subyearling release).
Used to back-calculate the broodtake from \code{n_yearling} and \code{n_subyearling}.}

\item{\code{Mocean_HOS}}{Either vector by age (length \code{maxage}) or an array with dimension \verb{[nsim, maxage, nyears+proyears]}.
Natural mortality of immature hatchery origin fish.
To replicate the SAR parameter of a stage-specific model, set \code{Mocean_HOS[a] = -log(SAR)} for the age class prior to maturation (and zero for all other ages).}

\item{\code{gamma}}{Numeric. The relative fecundity of hatchery origin spawners relative to natural origin spawners.}

\item{\code{pmax_NOB}}{Numeric. The maximum proportion of the natural origin escapement to be used as broodtake.}

\item{\code{ptarget_NOB}}{Numeric. The target proportion of the natural origin broodtake relative to the overall broodtake.}

\item{\code{phatchery}}{Numeric. The proportion of the hatchery origin escapement that return to the hatchery, for example, by removal from spawning grounds
or swim-in facilities. These fish are available for broodtake.}

\item{\code{premove_HOS}}{Numeric. The proportion of the hatchery origin escapement to be removed from the spawning grounds (in order to
ensure a high proportion of NOS). These fish are not available for broodtake. For example, a value less than one can represent
imperfect implementation of weir removal.}

\item{\code{fec_brood}}{Numeric. The fecundity of broodtake.}

\item{\code{fitness_type}}{Character. Fitness function for hatchery origin spawners. Either "Ford" or "none".}

\item{\code{theta}}{Vector length 2. The optimum phenotype value for the natural and hatchery environments.}

\item{\code{rel_loss}}{Vector length 3. The loss in fitness apportioned among the egg, fry, and smolt life stages. The three values should sum to 1.}

\item{\code{pbar_start}}{Vector length 2. The mean phenotype value in the the natural and hatchery population at the start of the projection.}

\item{\code{fitness_variance}}{Numeric. The variance of the phenotype in the population. Assumed identical between the natural and hatchery environments.}

\item{\code{selection_strength}}{Numeric. The ratio between the fitness standard deviation and the phenotype standard deviation.}

\item{\code{heritability}}{Numeric. The heritability of the phenotypic trait. Between 0-1.}

\item{\code{fitness_floor}}{Numeric. The minimum fitness value in the natural environment.}

\item{\code{u_preterminal}}{Numeric. The exploitation rate of the immature stock in the pre-terminal fishery.}

\item{\code{u_terminal}}{Numeric. The exploitation rate of the return in the terminal fishery.}

\item{\code{m}}{Numeric. The mark rate of hatchery origin fish, which affects fishery selectivity. Mark-selective fishing (MSF) is implemented if \code{m > 0}.}

\item{\code{release_mort}}{Vector length 2. The proportion of released fish that die after release, in the pre-terminal and terminal fishery.
Implemented to model mark-selective fishing. Not used if \code{m = 0}.}

\item{\code{vulPT}}{Vector length \code{maxage}. Vulnerability schedule (between 0-1) in the preterminal fishery.}

\item{\code{vulT}}{Vector length \code{maxage}. Vulnerability schedule (between 0-1) in the terminal fishery.}

\item{\code{HistSmolt}}{Vector by historical years (\code{nyears}) or a matrix by \verb{[nsim, nyears]}. The natural smolt production.}

\item{\code{HistSpawner}}{Array by \verb{[nsim, maxage, nyears]}. Total spawners (natural and hatchery origin combined) in the natural environment.}

\item{\code{HistN}}{Array by \verb{[nsim, maxage, nyears, 2]}. The abundance of immature fish at the beginning of the annual time step. The first array slice
corresponds to natural origin fish and the second array slice corresponds to hatchery origin fish.}

\item{\code{HistYearling}}{Vector by historical years. Hatchery releases (number of smolts).}

\item{\code{HistFPT}}{Vector by historical years (\code{nyears}) or an array by dimension \verb{[nsim, nyears, 2]}. The instantaneous fishing mortality in the preterminal fishery.
The first array slice corresponds to F for natural origin fish and the second array slice corresponds to hatchery origin fish.}

\item{\code{HistFT}}{Vector by historical years (\code{nyears}) or an array by dimension \verb{[nsim, nyears, 2]}. The instantaneous fishing mortality in the terminal fishery.
The first array slice corresponds to F for natural origin fish and the second array slice corresponds to hatchery origin fish.}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("SOM", Bio, Habitat, Hatchery, Harvest, Historical)}.
}

\keyword{classes}
